refactor(hunt): Modularize hunt component and tests
