+++ .gitignore
@@ -19,3 +19,5 @@ __pycache__
 __debug_bin*
 .venv
 html/so-ca.crt
+
+socdev.sh
