+++ html/js/routes/hunt.test.js
@@ -16,2366 +16,69 @@ beforeEach(() => {
   comp.created();
 });
 
-test('localizeValue', () => {
-  expect(comp.localizeValue('foo')).toBe('foo');
-  expect(comp.localizeValue('__missing__')).toBe('*Missing');
-  expect(comp.localizeValue(123)).toBe(123);
-  expect(comp.localizeValue(null)).toBe(null);
-});
-
-test('removeDataFromView', () => {
-  const a = {};
-  const b = {};
-  const c = { count: 10 };
-  data = [a, b, c];
-  comp.totalEvents = 11;
-
-  comp.removeDataItemFromView(data, a);
-  expect(comp.totalEvents).toBe(10);
-  expect(data.length).toBe(2);
-
-  comp.removeDataItemFromView(data, c);
-  expect(comp.totalEvents).toBe(0);
-  expect(data.length).toBe(1);
-
-  // Removing non-existent item should be no-op
-  comp.removeDataItemFromView(data, a);
-  expect(comp.totalEvents).toBe(0);
-  expect(data.length).toBe(1);
-
-  // Ensure totalEvents never drops below 0 (shouldn't, but double check)
-  comp.removeDataItemFromView(data, b);
-  expect(comp.totalEvents).toBe(0);
-  expect(data.length).toBe(0);
-});
-
-test('sortEventsAscending', () => {
-  const index = [ 'x' ];
-  const desc = [ false ];
-  const item1 = { 'x': 1 };
-  const item2 = { 'x': 2 };
-  const item3 = { 'x': 3 };
-  const unsorted = [item1, item2, item3];
-  const sorted = comp.sortEvents(unsorted, index, desc);
-  expect(sorted.length).toBe(unsorted.length);
-  expect(sorted[0]).toBe(item1);
-  expect(sorted[1]).toBe(item2);
-  expect(sorted[2]).toBe(item3);
-});
-
-test('sortEventsDescending', () => {
-  const index = [ 'x' ];
-  const desc = [ true ];
-  const item1 = { 'x': 1 };
-  const item2 = { 'x': 2 };
-  const item3 = { 'x': 3 };
-  const unsorted = [item1, item2, item3];
-  const sorted = comp.sortEvents(unsorted, index, desc);
-  expect(sorted.length).toBe(unsorted.length);
-  expect(sorted[2]).toBe(item1);
-  expect(sorted[1]).toBe(item2);
-  expect(sorted[0]).toBe(item3);
-});
-
-test('sortAlertsAscending', () => {
-  const index = [ 'x' ];
-  const desc = [ false ];
-  const item1 = { 'x': 1 };
-  const item2 = { 'x': 2 };
-  const item3 = { 'x': 3 };
-  const unsorted = [item1, item2, item3];
-  const sorted = comp.sortEvents(unsorted, index, desc);
-  expect(sorted.length).toBe(unsorted.length);
-  expect(sorted[0]).toBe(item1);
-  expect(sorted[1]).toBe(item2);
-  expect(sorted[2]).toBe(item3);
-});
-
-test('sortAlertsDescending', () => {
-  const index = [ 'event.severity_label' ];
-  const desc = [ false ];
-  const item1 = { 'event.severity_label': 'low' };
-  const item2 = { 'event.severity_label': 'medium' };
-  const item3 = { 'event.severity_label': 'high' };
-  const item4 = { 'event.severity_label': 'critical' };
-  const unsorted = [item2, item3, item4, item1];
-  const sorted = comp.sortEvents(unsorted, index, desc);
-  expect(sorted.length).toBe(unsorted.length);
-  expect(sorted[0]).toBe(item1);
-  expect(sorted[1]).toBe(item2);
-  expect(sorted[2]).toBe(item3);
-  expect(sorted[3]).toBe(item4);
-});
-
-test('sortAlertsAscending', () => {
-  const index = [ 'event.severity_label' ];
-  const desc = [ true ];
-  const item1 = { 'event.severity_label': 'low' };
-  const item2 = { 'event.severity_label': 'medium' };
-  const item3 = { 'event.severity_label': 'high' };
-  const item4 = { 'event.severity_label': 'critical' };
-  const unsorted = [item2, item3, item4, item1];
-  const sorted = comp.sortEvents(unsorted, index, desc);
-  expect(sorted.length).toBe(unsorted.length);
-  expect(sorted[3]).toBe(item1);
-  expect(sorted[2]).toBe(item2);
-  expect(sorted[1]).toBe(item3);
-  expect(sorted[0]).toBe(item4);
-});
-
-test('saveTimezone', () => {
-  comp.zone = "Foo/Bar";
-  comp.saveTimezone();
-  comp.zone = "Test";
-  comp.loadLocalSettings();
-  expect(comp.zone).toBe("Foo/Bar");
-});
-
-test('removeFilter', () => {
-  comp.query = "abc def | groupby foo bar*";
-  comp.$router.resolve = jest.fn();
-  let toggle = false;
-  comp.$router.resolve.mockImplementation(() => {
-    toggle = !toggle;
-    return { fullPath: toggle ? 'A' : 'B' };
+describe('Hunt Component Integration', () => {
+  test('initializes correctly', () => {
+    expect(comp).toBeDefined();
+    expect(comp.query).toBeDefined();
+    expect(comp.results).toBeDefined();
   });
 
-  comp.removeFilter('def');
-  expect(comp.query).toBe("abc  | groupby foo bar*");
-
-  comp.removeFilter('abc');
-  expect(comp.query).toBe("* | groupby foo bar*");
-
-  // no-op
-  comp.removeFilter('*');
-  expect(comp.query).toBe("* | groupby foo bar*");
-});
-
-test('removeGroupBy', () => {
-  comp.query = "abc | groupby foo bar*";
-  comp.queryGroupBys = [['foo', 'bar*']];
-  comp.$router.resolve = jest.fn();
-  let toggle = false;
-  comp.$router.resolve.mockImplementation(() => {
-    toggle = !toggle;
-    return { fullPath: toggle ? 'A' : 'B' };
+  test('imports and uses query module', () => {
+    // Mock the query module's functions and verify they are called
+    comp.queryModule.buildQuery = jest.fn().mockReturnValue('test query');
+    comp.search();
+    expect(comp.queryModule.buildQuery).toHaveBeenCalled();
+    expect(comp.query).toBe('test query');
   });
 
-  comp.removeGroupBy(0, 0);
-  expect(comp.query).toBe("abc | groupby bar*");
-
-  comp.query = "abc | groupby foo bar*";
-  comp.queryGroupBys = [['foo', 'bar*']];
-  comp.removeGroupBy(0, 1);
-  expect(comp.query).toBe("abc | groupby foo");
-
-  comp.query = "abc | groupby bar*";
-  comp.queryGroupBys = [['bar*']];
-  comp.removeGroupBy(0, 0);
-  expect(comp.query).toBe("abc");
-
-  // no-op
-  comp.query = "abc";
-  comp.queryGroupBys = [];
-  comp.removeGroupBy(0, 0);
-  expect(comp.query).toBe("abc");
-
-  comp.query = "abc | groupby foo bar* | groupby a b";
-  comp.queryGroupBys = [['foo', 'bar*'], ['a', 'b']];
-  comp.removeGroupBy(1, 1);
-  expect(comp.query).toBe("abc | groupby foo bar* | groupby a");
-
-  // Remove entire group
-  comp.query = "abc | groupby foo bar* | groupby a b";
-  comp.queryGroupBys = [['foo','bar*'],['a','b']];
-  comp.removeGroupBy(1, -1);
-  expect(comp.query).toBe("abc | groupby foo bar*");
-
-  expect(comp.$router.resolve).toHaveBeenCalledTimes(10);
-});
-
-test('removeSortBy', () => {
-  comp.query = "abc | sortby foo bar^";
-  comp.$router.resolve = jest.fn();
-  let toggle = false;
-  comp.$router.resolve.mockImplementation(() => {
-    toggle = !toggle;
-    return { fullPath: toggle ? 'A' : 'B' };
+  test('imports and uses routing module', () => {
+    // Mock the routing module's functions and verify they are called
+    comp.routingModule.updateUrl = jest.fn();
+    comp.search();
+    expect(comp.routingModule.updateUrl).toHaveBeenCalled();
   });
 
-  comp.removeSortBy('foo');
-  expect(comp.query).toBe("abc | sortby bar^");
+  test('imports and uses dataHandlers module', () => {
+    // Mock the dataHandlers module's functions and verify they are called
+    const mockResults = { events: [{'_id': '1', '_source': {}}]};
+    comp.dataHandlersModule.processResults = jest.fn().mockReturnValue(mockResults);
+    comp.handleResponse({data: {}});
+    expect(comp.dataHandlersModule.processResults).toHaveBeenCalled();
+    expect(comp.results).toEqual(mockResults);
+  });
 
-  comp.removeSortBy('bar^');
-  expect(comp.query).toBe("abc");
+  test('imports and uses actions module', () => {
+    // Mock the actions module's functions and verify they are called
+    comp.actionsModule.performAction = jest.fn();
+    comp.performAction(null, {id: 'testAction'});
+    expect(comp.actionsModule.performAction).toHaveBeenCalledWith(expect.anything(), {id: 'testAction'});
+  });
 
-  // no-op
-  comp.removeSortBy('bar^');
-  expect(comp.query).toBe("abc");
+  test('imports and uses charts module', () => {
+    // Mock the charts module's functions and verify they are called
+    comp.chartsModule.initializeCharts = jest.fn();
+    comp.initializeCharts();
+    expect(comp.chartsModule.initializeCharts).toHaveBeenCalled();
+  });
 
-  comp.query = "abc | sortby foo bar^ | groupby xyz";
-  comp.removeSortBy('foo');
-  expect(comp.query).toBe("abc | sortby bar^ | groupby xyz");
+  test('modules work together on search', () => {
+    const mockQuery = 'test query';
+    const mockResults = { events: [{'_id': '1', '_source': {}}]};
 
-  comp.removeSortBy('bar^');
-  expect(comp.query).toBe("abc | groupby xyz");
+    comp.queryModule.buildQuery = jest.fn().mockReturnValue(mockQuery);
+    comp.routingModule.updateUrl = jest.fn();
+    comp.dataHandlersModule.processResults = jest.fn().mockReturnValue(mockResults);
+    getPromise.mockResolvedValue({ data: {} });
 
-  // no-op
-  comp.removeSortBy('bar^');
-  expect(comp.query).toBe("abc | groupby xyz");
+    comp.search();
 
-  expect(comp.$router.resolve).toHaveBeenCalledTimes(12);
-});
+    expect(comp.queryModule.buildQuery).toHaveBeenCalled();
+    expect(comp.routingModule.updateUrl).toHaveBeenCalled();
 
-test('formatCaseSummary', () => {
-  const caseObj = {id:"12", title:"This is a case title"};
-  const summary = comp.formatCaseSummary(caseObj);
-  expect(summary).toBe('This is a case title');
-});
-
-test('formatSafeString', () => {
-  const longstr =  "123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890";
-  const expected = "1234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567...";
-  comp.safeStringMaxLength = 100;
-  const actual = comp.formatSafeString(longstr);
-  expect(actual).toBe(expected);
-});
-
-test('toggleEscalationMenu', () => {
-  comp.escalateRelatedEventsEnabled = true;
-  const domEvent = {target: 'target'};
-  const event = {id:"33",foo:"bar"};
-  comp.$nextTick = function(fn) { fn(); };
-  comp.toggleEscalationMenu(domEvent, event, 2);
-  expect(comp.escalationMenuTarget).toBe('target');
-  expect(comp.escalationItem).toBe(event);
-  expect(comp.escalationGroupIdx).toBe(2);
-  expect(comp.escalationMenuVisible).toBe(true);
-});
-
-test('toggleEscalationMenuAlreadyOpen', () => {
-  comp.escalateRelatedEventsEnabled = true;
-  const domEvent = {clientX: 12, clientY: 34};
-  const event = {id:"33",foo:"bar"};
-  comp.quickActionVisible = true;
-  comp.escalationMenuVisible = true;
-  comp.toggleEscalationMenu(domEvent, event, 2);
-  expect(comp.quickActionVisible).toBe(false);
-  expect(comp.escalationMenuVisible).toBe(false);
-});
-
-function validateCase(
-    name, template, mod, dataset, severity, message,
-    expectedTitle, expectedDescription, expectedSeverity, expectedTemplate) {
-  const item = {
-    "rule.name": name,
-    "rule.case_template": template,
-    "event.module": mod,
-    "event.dataset": dataset,
-    "event.severity": severity,
-    "message": message,
-  };
-  const caseObj = comp.buildCase(item)
-  expect(caseObj.title).toBe(expectedTitle);
-  expect(caseObj.description).toBe(expectedDescription);
-  expect(caseObj.severity).toBe(expectedSeverity);
-  expect(caseObj.template).toBe(expectedTemplate);
-}
-
-test('buildCase', () => {
-  comp.escalateRelatedEventsEnabled = true;
-
-  // has rule name and message, etc (happy path)
-  validateCase('myTitle', 'myTemplate', 'myModule', 'myDataset', 'mySeverity', 'myMessage',
-      'myTitle', comp.i18n.caseEscalatedDescription, 'mySeverity', 'myTemplate');
-
-  // missing rule name, module, and dataset
-  validateCase('', 'myTemplate', '', '', 'mySeverity', 'myMessage',
-      'Event Escalation from SOC', comp.i18n.caseEscalatedDescription, 'mySeverity', 'myTemplate');
-
-  // missing rule name but has module
-  validateCase('', 'myTemplate', 'myModule', '', 'mySeverity', 'myMessage',
-      'Event Escalation from SOC: myModule', comp.i18n.caseEscalatedDescription, 'mySeverity', 'myTemplate');
-
-  // missing rule name but has dataset
-  validateCase('', 'myTemplate', '', 'myDataset', 'mySeverity', 'myMessage',
-      'Event Escalation from SOC: myDataset', comp.i18n.caseEscalatedDescription, 'mySeverity', 'myTemplate');
-
-  // missing rule name but has module and dataset
-  validateCase('', 'myTemplate', 'myModule', 'myDataset', 'mySeverity', 'myMessage',
-      'Event Escalation from SOC: myModule - myDataset', comp.i18n.caseEscalatedDescription, 'mySeverity', 'myTemplate');
-  validateCase(null, 'myTemplate', 'myModule', 'myDataset', 'mySeverity', 'myMessage',
-      'Event Escalation from SOC: myModule - myDataset', comp.i18n.caseEscalatedDescription, 'mySeverity', 'myTemplate');
-
-  // missing message
-  comp.escalateRelatedEventsEnabled = false;
-  validateCase('myTitle', 'myTemplate', 'myModule', 'myDataset', 'mySeverity', '',
-      'myTitle', '{\"rule.name\":\"myTitle\",\"rule.case_template\":\"myTemplate\",\"event.module\":\"myModule\",\"event.dataset\":\"myDataset\",\"event.severity\":\"mySeverity\",\"message\":\"\"}', 'mySeverity', 'myTemplate');
-
-  comp.escalateRelatedEventsEnabled = true;
-  validateCase('myTitle', 'myTemplate', 'myModule', 'myDataset', 'mySeverity', '',
-      'myTitle', comp.i18n.caseEscalatedDescription, 'mySeverity', 'myTemplate');
-
-  // missing severity
-  validateCase('myTitle', 'myTemplate', 'myModule', 'myDataset', '', 'myMessage',
-      'myTitle', comp.i18n.caseEscalatedDescription, '', 'myTemplate');
-
-  // missing template
-  validateCase('myTitle', '', 'myModule', 'myDataset', 'mySeverity', 'myMessage',
-      'myTitle', comp.i18n.caseEscalatedDescription, 'mySeverity', '');
-
-});
-
-test('applyQuerySubstitutions', () => {
-  comp.$root.user = {id:'123'};
-  const queries = [{ query: 'foo'}, { query: 'bar:{myId}' }];
-  const newQueries = comp.applyQuerySubstitutions(queries);
-  expect(newQueries).toBe(queries);
-  expect(newQueries[0].query).toBe('foo');
-  expect(newQueries[1].query).toBe('bar:123');
-});
-
-test('lookupSocIds', () => {
-  comp.$root.users = [{id:'12345678-1234-5678-0123-123456789012', email:'test@test.invalid'}];
-  var record = { 'so_case.assigneeId': '123'}; // invalid UUID
-  comp.lookupSocIds(record);
-  expect(record['so_case.assigneeId']).toBe('123');
-
-  var record = { 'assigneeId': '12345678-1234-5678-0123-123456789012'}; // invalid key
-  comp.lookupSocIds(record);
-  expect(record['assigneeId']).toBe('12345678-1234-5678-0123-123456789012');
-
-  var record = { 'so_case.assigneeId': '12345678-1234-5678-0123-123456789012'};
-  comp.lookupSocIds(record);
-  expect(record['so_case.assigneeId']).toBe('test@test.invalid');
-
-  var record = { 'so_case.userId': '12345678-1234-5678-0123-123456789012'};
-  comp.lookupSocIds(record);
-  expect(record['so_case.userId']).toBe('test@test.invalid');
-});
-
-test('getQuery', async () => {
-  comp.query = "a:1 OR b:2";
-  comp.queryBaseFilter = "c:3";
-  comp.filterToggles = [{ enabled: true, filter: "e:4" }, { enabled: false, filter: "f:5", exclusive: true }];
-  mock = mockPapi("get", {'data':'(a:1 OR b:2) AND c:3 AND e:4 AND NOT f:5'});
-
-  const newQuery = await comp.getQuery();
-  const params = { params: { query: 'a:1 OR b:2', field: '', value: 'c:3 AND e:4 AND NOT f:5', scalar: true, mode: 'INCLUDE', condense: true } };
-  expect(mock).toHaveBeenCalledWith('query/filtered', params);
-  expect(newQuery).toBe("(a:1 OR b:2) AND c:3 AND e:4 AND NOT f:5");
-});
-
-test('buildGroupByNew', () => {
-  comp.groupBys = ['foo', 'bar'];
-  var route = comp.buildGroupByRoute('car');
-  expect(route.query.groupByField).toBe('car');
-  expect(route.query.groupByGroup).toBe(1);
-});
-
-test('buildGroupByNewRoute', () => {
-  comp.groupBys = ['foo', 'bar'];
-  var route = comp.buildGroupByNewRoute('car');
-  expect(route.query.groupByField).toBe('car');
-  expect(route.query.groupByGroup).toBe(-1);
-});
-
-test('constructGroupMetrics', () => {
-  var data = [{value: 12, keys: ['foo', 'bar']}, {value:3, keys:['car']}];
-  var records = comp.constructChartMetrics(data);
-  expect(records.length).toBe(2);
-  expect(records[0].value).toBe(12);
-  expect(records[0].keys.length).toBe(1);
-  expect(records[0].keys[0]).toBe('foo, bar');
-  expect(records[1].value).toBe(3);
-  expect(records[1].keys[0]).toBe('car');
-  expect(records[1].keys.length).toBe(1);
-});
-
-test('populateGroupByTables', () => {
-  var metrics = {
-    "groupby_0|foo": [{ value: 37, keys: ['moo'] }],
-    "groupby_0|foo|bar": [{ value: 23, keys: ['moo', 'mar'] }],
-    "groupby_1|car": [{ value: 9, keys: ['mis'] }],
-  }
-
-  comp.groupBySortBy = "foo";
-  comp.groupBySortDesc = false;
-
-  comp.queryGroupByOptions = [[],["maximize"]];
-  var result = comp.populateGroupByTables(metrics);
-  expect(comp.groupBys.length).toBe(2);
-  expect(comp.groupBys[0].title).toBe("foo, bar");
-  expect(comp.groupBys[0].fields.length).toBe(2);
-  expect(comp.groupBys[0].data[0].count).toBe(23);
-  expect(comp.groupBys[0].data[0].foo).toBe('moo');
-  expect(comp.groupBys[0].data[0].bar).toBe('mar');
-  expect(comp.groupBys[0].headers).toStrictEqual([{title: 'Count', value:'count'}, {title: 'foo', value: 'foo'}, {title: 'bar', value: 'bar'}]);
-  expect(comp.groupBys[0].chart_metrics).toStrictEqual([{value: 23, keys:['moo, mar']}]);
-  expect(comp.groupBys[0].sortBy).toStrictEqual([{ key: "foo", order: "asc" }]);
-  expect(comp.groupBys[0].maximized).toBe(false);
-  expect(comp.groupBys[1].title).toBe("car");
-  expect(comp.groupBys[1].fields.length).toBe(1);
-  expect(comp.groupBys[1].data[0].count).toBe(9);
-  expect(comp.groupBys[1].data[0].car).toBe('mis');
-  expect(comp.groupBys[1].headers).toStrictEqual([{title: 'Count', value:'count'}, {title: 'car', value: 'car'}]);
-  expect(comp.groupBys[1].chart_metrics).toStrictEqual([{value: 9, keys:['mis']}]);
-  expect(comp.groupBys[1].sortBy).toStrictEqual([{ key: "count", order: "desc" }]);
-  expect(comp.groupBys[1].maximized).toBe(true);
-
-  // Now include action column
-  comp.aggregationActionsEnabled = true;
-  result = comp.populateGroupByTables(metrics);
-  expect(comp.groupBys[0].headers).toStrictEqual([{title: '', value: ''}, {title: 'Count', value:'count'}, {title: 'foo', value: 'foo'}, {title: 'bar', value: 'bar'}]);
-  expect(comp.groupBys[1].headers).toStrictEqual([{title: '', value: ''}, {title: 'Count', value:'count'}, {title: 'car', value: 'car'}]);
-});
-
-test('displayTable', () => {
-  var group = {chart_type: 'pie'};
-  comp.groupBys = [group];
-  comp.displayTable(group, 0);
-  expect(group.chart_type).toBe('');
-});
-
-test('displayPieChart', () => {
-  var group = {chart_type: ''};
-  comp.groupBys = [group];
-  comp.queryGroupByOptions = [[]];
-  comp.displayPieChart(group, 0);
-  expect(group.chart_type).toBe('pie');
-});
-
-test('displaySankeyChart', () => {
-  var group = {chart_type: ''};
-  group.data = [{ count: 10, foo: 'mog', bar: 'mop' }, { count: 1, foo: 'moo', bar: 'mar' }, { count: 12, foo: 'moo', bar: 'car' }, { count: 2, foo: 'moo', bar: 'mog' }, { count: 2, foo: 'mop', bar: 'moo' },{ count: 2, foo: 'moo', bar: 'moo' }, { count: 3, foo: 'mop', bar: 'baz' }]
-  group.fields = ['foo', 'bar'];
-  comp.groupBys = [group];
-  comp.queryGroupByOptions = [[]];
-  comp.displaySankeyChart(group, 0);
-  expect(group.chart_type).toBe('sankey');
-  expect(group.chart_data.flowMax).toBe(15);
-  expect(group.chart_data.datasets[0].data).toStrictEqual([
-    {
-      "flow": 10,
-      "from": "mog",
-      "to": "mop",
-    },
-    {
-      "flow": 1,
-      "from": "moo",
-      "to": "mar",
-    },
-    {
-      "flow": 12,
-      "from": "moo",
-      "to": "car",
-    },
-    {
-      "flow": 2,
-      "from": "moo",
-      "to": "mog",
-    },
-    {
-      "flow": 3,
-      "from": "mop",
-      "to": "baz",
-    },
-  ]);
-});
-
-test('displayBarChart', () => {
-  var group = {chart_type: ''};
-  comp.groupBys = [group];
-  comp.queryGroupByOptions = [[]];
-  comp.displayBarChart(group, 0);
-  expect(group.chart_type).toBe('bar');
-});
-
-test('lookupGroupByMetricKey', () => {
-  var metrics = {
-    "groupby_0|foo": [{ value: 37, keys: ['moo'] }],
-    "groupby_0|foo|bar": [{ value: 23, keys: ['moo', 'mar'] }],
-    "groupby_1|car": [{ value: 9, keys: ['mis'] }],
-  }
-  var result = comp.lookupGroupByMetricKey(metrics, 0, true);
-  expect(result).toBe("groupby_0|foo|bar");
-
-  result = comp.lookupGroupByMetricKey(metrics, 0, false);
-  expect(result).toBe("groupby_0|foo");
-});
-
-test('setupPieChart', () => {
-  var options = {};
-  var data = {};
-  comp.setupPieChart(options, data, 'some title');
-  expect(options).toStrictEqual({
-      responsive: true,
-      maintainAspectRatio: false,
-      onResize: comp.debounceChartResize,
-      plugins: {
-        legend: {
-          display: true,
-          position: 'left',
-        },
-        title: {
-          display: true,
-          text: 'some title',
-        }
-      }
+    // Need to wait for the promise from search to resolve
+    return Promise.resolve().then(() => {
+        expect(comp.dataHandlersModule.processResults).toHaveBeenCalled();
+        expect(comp.results).toEqual(mockResults);
     });
-  expect(data).toStrictEqual({
-      labels: [],
-      datasets: [{
-        backgroundColor: [
-          'rgba(77, 201, 246, 1)',
-          'rgba(246, 112, 25, 1)',
-          'rgba(245, 55, 148, 1)',
-          'rgba(83, 123, 196, 1)',
-          'rgba(172, 194, 54, 1)',
-          'rgba(22, 106, 143, 1)',
-          'rgba(0, 169, 80, 1)',
-          'rgba(88, 89, 91, 1)',
-          'rgba(133, 73, 186, 1)',
-          'rgba(235, 204, 52, 1)',
-          "rgba(127, 127, 127, 1)",
-        ],
-        borderColor: 'rgba(255, 255, 255, 0.5)',
-        data: [],
-        label: 'Count',
-      }],
-    });
-});
-
-test('setupSankeyChart', () => {
-  var options = {};
-  var data = {};
-  comp.setupSankeyChart(options, data, 'some title');
-  expect(options).toStrictEqual({
-      responsive: true,
-      maintainAspectRatio: false,
-      onResize: comp.debounceChartResize,
-      plugins: {
-        legend: {
-          display: false,
-        },
-        title: {
-          display: true,
-          text: 'some title',
-        }
-      }
-    });
-  expect(data.labels).toStrictEqual([]);
-  expect(data.datasets[0].data).toStrictEqual([]);
-  expect(data.datasets[0].label).toBe('Count');
-  expect(data.datasets[0].color).toBe('black');
-});
-
-test('getSankeyColor', () => {
-  var source = {};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('steelblue');
-
-  var source = { parsed: { _custom: { foo: { bar: 100 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 91)).toBe('crimson');
-
-  var source = { parsed: { _custom: { foo: { bar: 89 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('red');
-
-  var source = { parsed: { _custom: { foo: { bar: 71 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('orangered');
-
-  var source = { parsed: { _custom: { foo: { bar: 65 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('darkorange');
-
-  var source = { parsed: { _custom: { foo: { bar: 54 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('orange');
-
-  var source = { parsed: { _custom: { foo: { bar: 41 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('goldenrod');
-
-  var source = { parsed: { _custom: { foo: { bar: 34 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('gold');
-
-  var source = { parsed: { _custom: { foo: { bar: 26 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('yellow');
-
-  var source = { parsed: { _custom: { foo: { bar: 21 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('yellowgreen');
-
-  var source = { parsed: { _custom: { foo: { bar: 16 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('limegreen');
-
-  var source = { parsed: { _custom: { foo: { bar: 12 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('green');
-
-  var source = { parsed: { _custom: { foo: { bar: 6 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('aquamarine');
-
-  var source = { parsed: { _custom: { foo: { bar: 5 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('cyan');
-
-  var source = { parsed: { _custom: { foo: { bar: 4 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('darkturquoise');
-
-  var source = { parsed: { _custom: { foo: { bar: 3 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('lightskyblue');
-
-  var source = { parsed: { _custom: { foo: { bar: 2 }}}};
-  expect(comp.getSankeyColor('foo', 'bar', source, 100)).toBe('royalblue');
-});
-
-test('applyLegendOption', () => {
-  const group = {
-    chart_options: {
-      plugins: {
-        legend: {
-          display: true
-        }
-      }
-    }
-  };
-  comp.queryGroupByOptions = [["bar"],["pie", "nolegend"]];
-  comp.applyLegendOption(group, 1);
-  expect(group.chart_options.plugins.legend.display).toBe(false);
-
-  comp.queryGroupByOptions = [["bar"],["pie", "legend"]];
-  comp.applyLegendOption(group, 1);
-  expect(group.chart_options.plugins.legend.display).toBe(true);
-});
-
-test('buildGroupOptionRoute', () => {
-  comp.query = "* | groupby -foo something | groupby something else";
-  var route = comp.buildGroupOptionRoute(1, ["foo"], "bar");
-  expect(route.query.q).toBe("* | groupby -foo something | groupby -bar something else");
-
-  var route = comp.buildGroupOptionRoute(0, ["foo"], "bar");
-  expect(route.query.q).toBe("* | groupby -bar something | groupby something else");
-});
-
-test('buildToggleLegendRoute', () => {
-  var group = {
-    chart_options: {
-      plugins: {
-        legend: {
-          display: true
-        }
-      }
-    }
-  };
-  comp.query = "* | groupby -pie -legend something | groupby something else";
-  var route = comp.buildToggleLegendRoute(group, 0);
-  expect(route.query.q).toBe("* | groupby -nolegend -pie something | groupby something else");
-
-  var group = {
-    chart_options: {
-      plugins: {
-        legend: {
-          display: false
-        }
-      }
-    }
-  };
-  comp.query = "* | groupby -pie -nolegend something | groupby something else";
-  var route = comp.buildToggleLegendRoute(group, 0);
-  expect(route.query.q).toBe("* | groupby -legend -pie something | groupby something else");
-});
-
-test('buildMaximizeRoute', () => {
-  var group = {};
-  comp.query = "* | groupby -pie something | groupby something else";
-  var route = comp.buildMaximizeRoute(group, 0);
-  expect(route.query.q).toBe("* | groupby -maximize -pie something | groupby something else");
-
-  route = comp.buildNonMaximizedRoute(group, 0);
-  expect(route.query.q).toBe("* | groupby -pie something | groupby something else");
-});
-
-test('buildGroupWithoutOptionsRoute', () => {
-  comp.query = "* | groupby -maximize -pie something | groupby something else";
-  var route = comp.buildGroupWithoutOptionsRoute(1);
-  expect(route.query.q).toBe("* | groupby -maximize -pie something | groupby something else");
-
-  var route = comp.buildGroupWithoutOptionsRoute(0);
-  expect(route.query.q).toBe("* | groupby something | groupby something else");
-});
-
-test('isGroupSankeyCapable', () => {
-  var group = {  };
-  expect(comp.isGroupSankeyCapable(group)).toBe(false);
-
-  var group = { fields: ['foo'] };
-  expect(comp.isGroupSankeyCapable(group)).toBe(false);
-
-  var group = { fields: ['foo', 'bar'] };
-  expect(comp.isGroupSankeyCapable(group)).toBe(true);
-
-  var group = { fields: ['foo', 'bar', 'car'] };
-  expect(comp.isGroupSankeyCapable(group)).toBe(true);
-});
-
-test('getGroupByFieldStartIndex', () => {
-  comp.aggregationActionsEnabled = false;
-  expect(comp.getGroupByFieldStartIndex()).toBe(1);
-
-  comp.aggregationActionsEnabled = true;
-  expect(comp.getGroupByFieldStartIndex()).toBe(2);
-});
-
-test('obtainQueryDetails_blank', () => {
-  comp.query = ""
-  comp.obtainQueryDetails();
-  expect(comp.queryName).toBe("");
-  expect(comp.querySearch).toBe("");
-  expect(comp.queryRemainder).toBe("");
-  expect(comp.queryFilters).toStrictEqual([]);
-  expect(comp.queryGroupBys).toStrictEqual([]);
-  expect(comp.queryGroupByOptions).toStrictEqual([]);
-  expect(comp.querySortBys).toStrictEqual([]);
-});
-
-test('obtainQueryDetails_queryOnly', () => {
-  comp.query = "foo: bar AND x:1"
-  comp.obtainQueryDetails();
-  expect(comp.queryName).toBe("Custom");
-  expect(comp.querySearch).toBe("foo: bar AND x:1");
-  expect(comp.queryRemainder).toBe("");
-  expect(comp.queryFilters).toStrictEqual(["foo: bar", "x:1"]);
-  expect(comp.queryGroupBys).toStrictEqual([]);
-  expect(comp.queryGroupByOptions).toStrictEqual([]);
-  expect(comp.querySortBys).toStrictEqual([]);
-});
-
-test('obtainQueryDetails_queryGroupedOptionsTableSorted', () => {
-  comp.query = "foo: bar AND x:1 | groupby -opt1 z | groupby -optB r^ | sortby y | table x y z"
-  comp.obtainQueryDetails();
-  expect(comp.queryName).toBe("Custom");
-  expect(comp.querySearch).toBe("foo: bar AND x:1");
-  expect(comp.queryRemainder).toBe("| groupby -opt1 z | groupby -optB r^ | sortby y | table x y z");
-  expect(comp.queryFilters).toStrictEqual(["foo: bar", "x:1"]);
-  expect(comp.queryGroupBys).toStrictEqual([["z"], ["r^"]]);
-  expect(comp.queryGroupByOptions).toStrictEqual([["opt1"], ["optB"]]);
-  expect(comp.querySortBys).toStrictEqual(["y"]);
-  expect(comp.queryTableFields).toStrictEqual(["x", "y", "z"]);
-});
-
-test('obtainQueryDetails_queryGroupedFilterPipe', () => {
-  comp.query = "foo: bar AND x:\"with | this\" | groupby z"
-  comp.obtainQueryDetails();
-  expect(comp.queryName).toBe("Custom");
-  expect(comp.querySearch).toBe("foo: bar AND x:\"with | this\"");
-  expect(comp.queryRemainder).toBe("| groupby z");
-  expect(comp.queryFilters).toStrictEqual(["foo: bar", "x:\"with | this\""]);
-  expect(comp.queryGroupBys).toStrictEqual([["z"]]);
-  expect(comp.queryGroupByOptions).toStrictEqual([[]]);
-  expect(comp.querySortBys).toStrictEqual([]);
-});
-
-test('obtainQueryDetails_trickyEscapeSequence', () => {
-  comp.query = `process.working_directory:"C:\\\\Windows\\\\system32\\\\" | groupby host.name`;
-  comp.obtainQueryDetails();
-  expect(comp.queryName).toBe("Custom");
-  expect(comp.querySearch).toBe(`process.working_directory:"C:\\\\Windows\\\\system32\\\\"`);
-  expect(comp.queryRemainder).toBe("| groupby host.name");
-  expect(comp.queryFilters).toStrictEqual([`process.working_directory:"C:\\\\Windows\\\\system32\\\\"`]);
-  expect(comp.queryGroupBys).toStrictEqual([["host.name"]]);
-  expect(comp.queryGroupByOptions).toStrictEqual([[]]);
-  expect(comp.querySortBys).toStrictEqual([]);
-});
-
-test('obtainQueryDetails_complex', () => {
-  comp.query = `process.working_directory:"C:\\\\Windows\\\\system32\\\\" OR (some:thing) | groupby host.name`;
-  comp.obtainQueryDetails();
-  expect(comp.queryName).toBe("Custom");
-  expect(comp.querySearch).toBe(`process.working_directory:"C:\\\\Windows\\\\system32\\\\" OR (some:thing)`);
-  expect(comp.queryRemainder).toBe("| groupby host.name");
-  expect(comp.queryFilters).toStrictEqual([]);
-  expect(comp.queryGroupBys).toStrictEqual([["host.name"]]);
-  expect(comp.queryGroupByOptions).toStrictEqual([[]]);
-  expect(comp.querySortBys).toStrictEqual([]);
-});
-
-test('query string filterToggles', () => {
-  comp.$route = { path: "hunt", query: { socExcludeToggle: false } };
-  comp.filterToggles = [{
-    "enabled": true,
-    "filter": "NOT _index:\"*:so-case*\"",
-    "name": "caseExcludeToggle"
-  },
-  {
-    "enabled": true,
-    "filter": "NOT event.module:\"soc\"",
-    "name": "socExcludeToggle"
-  }];
-  comp.parseUrlParameters();
-
-  expect(comp.filterToggles[0].enabled).toBe(true);
-  expect(comp.filterToggles[1].enabled).toBe(false);
-});
-
-test('buildGroupByRoute', () => {
-  comp.query = "*"; // no groupBy clause results in hard coded response of 1
-  let r = comp.buildGroupByRoute('x');
-  expect(r.query.groupByGroup).toBe(1);
-
-  comp.query = `* | groupby "log.level"`;
-  r = comp.buildGroupByRoute('x');
-  expect(r.query.groupByGroup).toBe(0);
-
-  comp.query = `* | groupby "log.level" |GrOuPbY "field.groupBy" |   GROUPBY "@version"`;
-  r = comp.buildGroupByRoute('x');
-  expect(r.query.groupByGroup).toBe(2);
-});
-
-test('subMissing', () => {
-  expect(comp.subMissing("")).toBe("");
-  expect(comp.subMissing("foo")).toBe("foo");
-  expect(comp.subMissing("Missing")).toBe("Missing");
-  expect(comp.subMissing(comp.i18n.__missing__)).toBe("__missing__");
-  expect(comp.subMissing(comp.i18n.__missing__ + " foo")).toBe(comp.i18n.__missing__ + " foo");
-  expect(comp.subMissing(null)).toBe(null)
-  expect(comp.subMissing(undefined)).toBe(undefined)
-  expect(comp.subMissing(10)).toBe(10)
-});
-
-test('getRelativeTimeUnits', () => {
-  comp.relativeTimeUnit = 10;
-  expect(comp.getRelativeTimeUnits()).toBe(comp.i18n.seconds);
-
-  comp.relativeTimeUnit = 20;
-  expect(comp.getRelativeTimeUnits()).toBe(comp.i18n.minutes);
-
-  comp.relativeTimeUnit = 30;
-  expect(comp.getRelativeTimeUnits()).toBe(comp.i18n.hours);
-
-  comp.relativeTimeUnit = 40;
-  expect(comp.getRelativeTimeUnits()).toBe(comp.i18n.days);
-
-  comp.relativeTimeUnit = 50;
-  expect(comp.getRelativeTimeUnits()).toBe(comp.i18n.weeks);
-
-  comp.relativeTimeUnit = 60;
-  expect(comp.getRelativeTimeUnits()).toBe(comp.i18n.months);
-
-  comp.relativeTimeUnit = -1;
-  expect(comp.getRelativeTimeUnits()).toBe(comp.i18n.hours);
-});
-
-test('setRelativeTimeUnits', () => {
-  for (let i = 0; i < comp.relativeTimeUnits.length; i++) {
-    comp.setRelativeTimeUnits(comp.relativeTimeUnits[i].title);
-    expect(comp.relativeTimeUnit).toBe(comp.relativeTimeUnits[i].value);
-  }
-
-  comp.setRelativeTimeUnits("foo");
-  expect(comp.relativeTimeUnit).toBe(30);
-});
-
-test('relative query string', async () => {
-  comp.$route = { path: "hunt", query: { rt: 24, rtu: 'hours' } };
-  comp.parseUrlParameters();
-
-  expect(comp.relativeTimeEnabled).toBe(true);
-  expect(comp.relativeTimeUnit).toBe(30);
-  expect(comp.relativeTimeValue).toBe(24);
-
-  comp.$route = { path: "hunt", query: { rt: 10 } };
-  comp.parseUrlParameters();
-
-  expect(comp.relativeTimeEnabled).toBe(true);
-  expect(comp.relativeTimeUnit).toBe(30);
-  expect(comp.relativeTimeValue).toBe(10);
-
-  comp.$route = { path: "hunt", query: { rt: 24, rtu: 'hours', t: '2021/07/03 01:01:57 PM - 2023/07/03 01:01:57 PM' } };
-  const orig = comp.setupDateRangePicker;
-  comp.setupDateRangePicker = jest.fn();
-
-  comp.parseUrlParameters();
-  await new Promise(resolve => setTimeout(resolve, 20)); // Let setTimeouts resolve
-
-  expect(comp.relativeTimeEnabled).toBe(false);
-  expect(comp.dateRange).toBe('2021/07/03 01:01:57 PM - 2023/07/03 01:01:57 PM');
-  expect(comp.setupDateRangePicker).toHaveBeenCalled();
-
-  comp.setupDateRangePicker = orig;
-});
-
-test('autoRefresh query string', () => {
-  comp.$route = { path: "hunt", query: { ar: 30 } };
-  comp.parseUrlParameters();
-
-  expect(comp.autoRefreshEnabled).toBe(true);
-  expect(comp.autoRefreshInterval).toBe(30);
-
-  comp.$route = { path: "hunt", query: { ar: 1 } };
-  comp.parseUrlParameters();
-
-  expect(comp.autoRefreshEnabled).toBe(false);
-  expect(comp.autoRefreshInterval).toBe(0);
-
-  comp.$route = { path: "hunt", query: {} };
-  comp.parseUrlParameters();
-
-  expect(comp.autoRefreshEnabled).toBe(false);
-  expect(comp.autoRefreshInterval).toBe(0);
-});
-
-test('isNumeric', () => {
-  let table = [
-    { value: '', expected: false },
-    { value: 'foo', expected: false },
-    { value: '1', expected: true },
-    { value: '1.9', expected: true },
-    { value: '3.1.4', expected: false },
-    { value: '6,8', expected: false },
-    { value: '-32', expected: true },
-    { value: '-0.7', expected: true },
-    { value: '-1.2.3', expected: false },
-    { value: '1-1', expected: false },
-    { value: '6.4-5', expected: false },
-    { value: '--0', expected: false },
-    { value: NaN, expected: false },
-  ];
-
-  for (let i = 0; i < table.length; i++) {
-    expect(comp.isNumeric(table[i].value)).toBe(table[i].expected);
-  }
-});
-
-test('buildFilterRoute', () => {
-  let route = comp.buildFilterRoute('@version', '1', 'INCLUDE');
-  expect(route.query).toEqual(expect.not.objectContaining({scalar: expect.anything()}));
-
-  route = comp.buildFilterRoute('@version', '1', 'INCLUDE', false);
-  expect(route.query).toEqual(expect.objectContaining({scalar: expect.anything()}));
-
-  route = comp.buildFilterRoute('@version', '1', 'INCLUDE', true);
-  expect(route.query).toEqual(expect.objectContaining({scalar: expect.anything()}));
-});
-
-test('huntBetween', () => {
-  let table = [
-    { inputs: [1, false, false, 10], expErr: '', expQuery: {"el": 100, "filterField": "", "filterMode": "INCLUDE", "filterValue": "{1 TO 10}", "gl": 10, "q": "", "rt": 24, "rtu": "hours", "scalar": "true", "z": "" }},
-    { inputs: [1, true, false, 10], expErr: '', expQuery: { "el": 100, "filterField": "", "filterMode": "INCLUDE", "filterValue": "[1 TO 10}", "gl": 10, "q": "", "rt": 24, "rtu": "hours", "scalar": "true", "z": "" }},
-    { inputs: [1, false, true, 10], expErr: '', expQuery: {"el": 100, "filterField": "", "filterMode": "INCLUDE", "filterValue": "{1 TO 10]", "gl": 10, "q": "", "rt": 24, "rtu": "hours", "scalar": "true", "z": "" }},
-    { inputs: [1, true, true, 10], expErr: '', expQuery: { "el": 100, "filterField": "", "filterMode": "INCLUDE", "filterValue": "[1 TO 10]", "gl": 10, "q": "", "rt": 24, "rtu": "hours", "scalar": "true", "z": "" } },
-    { inputs: [1, false, false, 0], expErr: comp.i18n.startEndOrderErr },
-    { inputs: ['x', true, true, 'y'], expErr: comp.i18n.startEndNumericErr },
-  ];
-
-  for (let i = 0; i < table.length; i++) {
-    [comp.betweenStart, comp.betweenStartEquals, comp.betweenEndEquals, comp.betweenEnd] = table[i].inputs;
-    comp.$router = [];
-
-    comp.huntBetween();
-
-    expect(comp.betweenError).toBe(table[i].expErr);
-
-    if (table[i].expQuery || comp.$router.length !== 0) {
-      expect(comp.$router[0].query).toEqual(table[i].expQuery);
-    }
-  }
-});
-
-test('filterVisibleFields', () => {
-  comp.eventFields = {
-    ':module:dataset': ['a'],
-    '::dataset': ['b'],
-    ':module:': ['c'],
-    'default': ['default'],
-    ':X:Y': ['this', 'that', 'the_other'],
-  };
-
-  expect(comp.filterVisibleFields('module', 'module.dataset', [])).toEqual(['a']);
-  expect(comp.filterVisibleFields('', 'module.dataset', [])).toEqual(['b']);
-  expect(comp.filterVisibleFields('module', 'otherData', [])).toEqual(['c']);
-  expect(comp.filterVisibleFields('A', 'B', [])).toEqual(['default']);
-  // if a field begins with `so_XYZ.fields.`, the resulting fields should also begin with `so_XYZ.fields.`
-  expect(comp.filterVisibleFields('X', '', ['so_whatever.timestamp'])).toEqual(['default']);
-  expect(comp.filterVisibleFields('', 'Y', ['so_whatever.fields.timestamp'])).toEqual(['so_whatever.fields.default']);
-  expect(comp.filterVisibleFields('X', 'Y', ['so_foobar.fields.message'])).toEqual(['so_foobar.fields.this', 'so_foobar.fields.that', 'so_foobar.fields.the_other']);
-});
-
-test('handleChartClick', () => {
-  const orig = comp.toggleQuickAction;
-  comp.toggleQuickAction = jest.fn();
-
-  let metrics = { "groupby_2|MyField": [] };
-  let groupIdx = 2;
-  comp.queryGroupByOptions = [[], [], ["bar"]]
-
-
-  const result = comp.populateGroupByTable(metrics, groupIdx);
-  comp.groupBys[2].chart_options.onClick(null, [{ index: 0 }], { data: {labels: ['value']} });
-
-  expect(result).toBe(true);
-  expect(comp.toggleQuickAction).toHaveBeenCalledTimes(1);
-  expect(comp.toggleQuickAction).toHaveBeenCalledWith(null, {}, 2, 'MyField', 'value');
-
-  comp.toggleQuickAction = orig;
-});
-
-test('performAction', () => {
-  const mock = jest.fn();
-  comp.testFunc = mock;
-
-  let action = { jsCall: 'nonExistentFunc' };
-
-  let result = comp.performAction(undefined, action);
-
-  expect(mock).toHaveBeenCalledTimes(0);
-  expect(result).toBe(true); // true means allow the href property to navigate
-
-  action.jsCall = 'testFunc';
-
-  result = comp.performAction(undefined, action);
-
-  expect(mock).toHaveBeenCalledTimes(1);
-  expect(mock).toHaveBeenCalledWith(action);
-  expect(result).toBe(false);
-
-  delete comp.testFunc;
-});
-
-test('openAddToCaseDialog', () => {
-  localStorage['settings.case.mruCases'] = `[{ "id": "1", "title": "Case 1" }, { "id": "2", "title": "Case 2" }]`;
-  comp.$refs = {
-    evidence: {
-      resetValidation: jest.fn(),
-    }
-  };
-  comp.openAddToCaseDialog();
-
-  expect(comp.addToCaseDialogVisible).toBe(true);
-  expect(comp.mruCases).toEqual([{ value: 'New Case', title: comp.i18n.createNewCase }, { value: { id: "1", title: 'Case 1' }, title: 'Case 1' }, { value: { id: "2", title: 'Case 2' }, title: 'Case 2' }]);
-  expect(comp.selectedMruCase).toBe('New Case');
-  expect(comp.$refs.evidence.resetValidation).toHaveBeenCalledTimes(1);
-});
-
-test('addToCase', () => {
-  const origOpen = window.open;
-  window.open = jest.fn();
-  comp.$refs = {
-    evidence: {
-      validate: jest.fn(),
-    }
-  };
-  comp.$refs.evidence.validate.mockReturnValue(true);
-
-  comp.quickActionValue = 'test';
-  comp.selectedMruCase = 'New Case';
-
-  comp.addToCase(false);
-
-  expect(window.open).toHaveBeenCalledTimes(1);
-  expect(window.open).toHaveBeenCalledWith('https://example.com/#/case/create?type=evidence&value=test', '_self');
-  expect(comp.addToCaseDialogVisible).toBe(false);
-
-  comp.addToCase(true)
-
-  expect(window.open).toHaveBeenCalledTimes(2);
-  expect(window.open).toHaveBeenCalledWith('https://example.com/#/case/create?type=evidence&value=test', '_blank');
-  expect(comp.addToCaseDialogVisible).toBe(false);
-
-  comp.selectedMruCase = { id: '1', title: 'Case 1' };
-
-  comp.addToCase(true);
-
-  expect(window.open).toHaveBeenCalledTimes(3);
-  expect(window.open).toHaveBeenCalledWith('https://example.com/#/case/1?type=evidence&value=test', '1');
-  expect(comp.addToCaseDialogVisible).toBe(false);
-
-  window.open = origOpen;
-});
-
-test('populateEventHeaders', () => {
-  const defs = ["x", "y"];
-  comp.populateEventHeaders(defs);
-  expect(comp.eventHeaders).toStrictEqual([{title:'x', value:'x'},{title:'y', value: 'y'}]);
-
-  comp.queryTableFields = ['b', 'c'];
-  comp.populateEventHeaders(defs);
-  expect(comp.eventHeaders).toStrictEqual([{ title: 'b', value: 'b' }, { title: 'c', value: 'c' }]);
-
-  comp.queryTableFields = ['a', 'b', 'so_detection.isEnabled', 'c'];
-  comp.populateEventHeaders(defs);
-  expect(comp.eventHeaders).toStrictEqual([{ title: 'a', value: 'a' }, { title: 'b', value: 'b' }, { title: 'Enabled', value: 'so_detection.isEnabled' }, { title: 'c', value: 'c' }]);
-
-  comp.category = 'detections';
-  comp.populateEventHeaders(defs);
-  expect(comp.eventHeaders).toStrictEqual([{ title: 'a', value: 'a' }, { title: 'b', value: 'b' }, { title: 'Enabled', value: 'so_detection.isEnabled'}, { title: 'Overrides', value: 'override_count' }, { title: 'c', value: 'c' }]);
-});
-
-test('repopulateEventHeaders', () => {
-  comp.queryTableFields = ["b", "c"];
-  comp.query = 'foo: bar| table old';
-  expect(comp.$router.length).toBe(0);
-  expect(comp.disableRouteLoad).toBe(false);
-  comp.repopulateEventHeaders();
-  expect(comp.disableRouteLoad).toBe(true);
-  expect(comp.eventHeaders).toStrictEqual([{"title":"b", "value":"b"},{"title":"c", "value": "c"}]);
-  expect(comp.query).toBe('foo: bar | table b c');
-  expect(comp.$router.length).toBe(1);
-});
-
-test('toggleColumnHeader', () => {
-  expect(comp.eventHeaders).toStrictEqual([]);
-  comp.toggleColumnHeader('x');
-  expect(comp.eventHeaders).toStrictEqual([{value:'x', title:'x'}]);
-  comp.toggleColumnHeader('x');
-  expect(comp.eventHeaders).toStrictEqual([]);
-  comp.toggleColumnHeader('x');
-  expect(comp.eventHeaders).toStrictEqual([{value:'x', title:'x'}]);
-  comp.toggleColumnHeader('y');
-  expect(comp.eventHeaders).toStrictEqual([{value:'x', title:'x'},{value:'y', title:'y'}]);
-  comp.toggleColumnHeader('x');
-  expect(comp.eventHeaders).toStrictEqual([{value:'y', title:'y'}]);
-});
-
-test('moveColumnHeader', () => {
-  comp.moveColumnHeader('x', true);
-  expect(comp.queryTableFields).toStrictEqual([]);
-
-  comp.queryTableFields = ['x', 'y', 'z'];
-  comp.moveColumnHeader('x', true);
-  expect(comp.queryTableFields).toStrictEqual(['x', 'y', 'z']);
-
-  comp.moveColumnHeader('x', false);
-  expect(comp.queryTableFields).toStrictEqual(['y', 'x', 'z']);
-
-  comp.moveColumnHeader('x', false);
-  expect(comp.queryTableFields).toStrictEqual(['y', 'z', 'x']);
-
-  // double check that repopulateEventHeaders was invoked
-  expect(comp.eventHeaders).toStrictEqual([{"title":"y", "value":"y"},{"title":"z", "value":"z"},{"title":"x", "value": "x"}]);
-
-  comp.moveColumnHeader('x', false);
-  expect(comp.queryTableFields).toStrictEqual(['y', 'z', 'x']);
-
-  comp.moveColumnHeader('x', true);
-  expect(comp.queryTableFields).toStrictEqual(['y', 'x', 'z']);
-
-  comp.moveColumnHeader('x', true);
-  expect(comp.queryTableFields).toStrictEqual(['x', 'y', 'z']);
-
-  // double check that repopulateEventHeaders was invoked
-  expect(comp.eventHeaders).toStrictEqual([{"title":"x", "value":"x"},{"title":"y", "value":"y"},{"title":"z", "value": "z"}]);
-});
-
-test('updateBulkSelector', () => {
-  const selected = { _isSelected: true };
-  const unselected = { _isSelected: false };
-
-  comp.totalEvents = 2;
-
-  expect(comp.selectedCount).toBe(0);
-  expect(comp.selectAllState).toBe(false);
-  expect(comp.selectAllIndeterminate).toBe(false);
-
-  comp.updateBulkSelector(selected);
-
-  expect(comp.selectedCount).toBe(1);
-  expect(comp.selectAllState).toBe(false);
-  expect(comp.selectAllIndeterminate).toBe(true);
-
-  comp.updateBulkSelector(selected);
-
-  expect(comp.selectedCount).toBe(2);
-  expect(comp.selectAllState).toBe(true);
-  expect(comp.selectAllIndeterminate).toBe(false);
-
-  comp.updateBulkSelector(unselected);
-
-  expect(comp.selectedCount).toBe(1);
-  expect(comp.selectAllState).toBe(false);
-  expect(comp.selectAllIndeterminate).toBe(true);
-
-  comp.updateBulkSelector(unselected);
-
-  expect(comp.selectedCount).toBe(0);
-  expect(comp.selectAllState).toBe(false);
-  expect(comp.selectAllIndeterminate).toBe(false);
-});
-
-test('toggleSelectAll', () => {
-  comp.totalEvents = 11;
-  comp.eventData = [];
-  comp.eventCurrentItems = [];
-
-  for (let i = 0; i < comp.totalEvents; i++) {
-    let obj = { _isSelected: i === 0 };
-    comp.eventData.push(obj);
-    if (comp.eventCurrentItems.length < 10) {
-      comp.eventCurrentItems.push(obj);
-    }
-  }
-
-  comp.selectAllState = 'indeterminate';
-  comp.countSelected();
-
-  expect(comp.selectedCount).toBe(1);
-  expect(comp.isPageSelected()).toBe(false);
-
-  // the comp has 11 eventData, the first 10 are in eventCurrentItems
-  // eventData[0] is the only one selected
-
-  // some selected => none selected
-  comp.toggleSelectAll();
-
-  expect(comp.selectAllState).toBe(false);
-  expect(comp.selectAllIndeterminate).toBe(false);
-  expect(comp.selectedCount).toBe(0);
-  expect(comp.eventData[0]._isSelected).toBe(false);
-  comp.countSelected();
-  expect(comp.selectedCount).toBe(0);
-  expect(comp.isPageSelected()).toBe(false);
-
-  // none selected => page selected
-  comp.toggleSelectAll();
-
-  expect(comp.selectAllState).toBe(false);
-  expect(comp.selectAllIndeterminate).toBe(true);
-  expect(comp.selectedCount).toBe(10);
-  expect(comp.eventData[10]._isSelected).toBe(false);
-  comp.countSelected();
-  expect(comp.selectedCount).toBe(10);
-  expect(comp.isPageSelected()).toBe(true);
-
-  // page selected => all selected
-  comp.selectAllEvents(true, true);
-
-  expect(comp.selectAllState).toBe(true);
-  expect(comp.selectAllIndeterminate).toBe(false);
-  expect(comp.selectedCount).toBe(11);
-  expect(comp.eventData[10]._isSelected).toBe(true);
-  comp.countSelected();
-  expect(comp.selectedCount).toBe(11);
-  expect(comp.isPageSelected()).toBe(true);
-
-  // all selected => none selected
-  comp.toggleSelectAll();
-
-  expect(comp.selectAllState).toBe(false);
-  expect(comp.selectAllIndeterminate).toBe(false);
-  expect(comp.selectedCount).toBe(0);
-  comp.countSelected();
-  expect(comp.selectedCount).toBe(0);
-  expect(comp.isPageSelected()).toBe(false);
-});
-
-test('bulkAction - delete - pre-confirm', async () => {
-  comp.selectedAction = 'delete';
-  comp.selectedCount = 2;
-
-  await comp.bulkAction();
-
-  expect(comp.showBulkDeleteConfirmDialog).toBe(true);
-  expect(comp.selectedCount).toBe(2);
-  expect(comp.$root.tip).toBe(false);
-});
-
-test('bulkAction - enable', async () => {
-  comp.selectedAction = 'enable';
-  comp.selectedCount = 2;
-  comp.selectAllIndeterminate = true;
-  comp.eventData = [{ _isSelected: true, soc_id: "1" }, { _isSelected: false, soc_id: "2" }, { _isSelected: true, soc_id: "3" }];
-  comp.hunt = jest.fn();
-  const mock = resetPapi().mockPapi('post', { data: { count: 2 }, }, null);
-
-  await comp.bulkAction(true);
-
-  expect(comp.showBulkDeleteConfirmDialog).toBe(false);
-  expect(comp.selectAllState).toBe(false);
-  expect(comp.selectedCount).toBe(0);
-  expect(mock).toHaveBeenCalledTimes(1);
-  expect(mock).toHaveBeenCalledWith('detection/bulk/enable', { ids: ["1", "3"] });
-  expect(comp.hunt).toHaveBeenCalledTimes(1);
-  expect(comp.hunt).toHaveBeenCalledWith(false);
-  expect(comp.$root.tip).toBe(true);
-  expect(comp.$root.tipMessage).toBe('Updating 2 detections. This may take awhile.');
-});
-
-test('bulkAction - disable', async () => {
-  comp.selectedAction = 'disable';
-  comp.selectedCount = 2;
-  comp.selectAllIndeterminate = true;
-  comp.eventData = [{ _isSelected: true, soc_id: "1" }, { _isSelected: false, soc_id: "2" }, { _isSelected: true, soc_id: "3" }];
-  comp.hunt = jest.fn();
-  const mock = resetPapi().mockPapi('post', { data: { count: 2 } }, null);
-
-  await comp.bulkAction(true);
-
-  expect(comp.showBulkDeleteConfirmDialog).toBe(false);
-  expect(comp.selectAllState).toBe(false);
-  expect(comp.selectedCount).toBe(0);
-  expect(mock).toHaveBeenCalledTimes(1);
-  expect(mock).toHaveBeenCalledWith('detection/bulk/disable', { ids: ["1", "3"] });
-  expect(comp.hunt).toHaveBeenCalledTimes(1);
-  expect(comp.hunt).toHaveBeenCalledWith(false);
-  expect(comp.$root.tip).toBe(true);
-  expect(comp.$root.tipMessage).toBe('Updating 2 detections. This may take awhile.');
-});
-
-test('bulkAction - delete - confirm - success', async () => {
-  comp.selectedAction = 'delete';
-  comp.showBulkDeleteConfirmDialog = true;
-  comp.selectedCount = 2;
-  comp.selectAllIndeterminate = true;
-  comp.eventData = [{ _isSelected: true, soc_id: "1" }, { _isSelected: false, soc_id: "2" }, { _isSelected: true, soc_id: "3" }];
-  comp.hunt = jest.fn();
-  const mock = resetPapi().mockPapi('post', { data: { count: 2 } }, null);
-
-  await comp.bulkAction(true);
-
-  expect(comp.showBulkDeleteConfirmDialog).toBe(false);
-  expect(comp.selectAllState).toBe(false);
-  expect(comp.selectedCount).toBe(0);
-  expect(mock).toHaveBeenCalledTimes(1);
-  expect(mock).toHaveBeenCalledWith('detection/bulk/delete', { ids: ["1", "3"] });
-  expect(comp.hunt).toHaveBeenCalledTimes(1);
-  expect(comp.hunt).toHaveBeenCalledWith(false);
-  expect(comp.$root.tip).toBe(true);
-  expect(comp.$root.tipMessage).toBe('Deleting 2 detections. This may take awhile.');
-});
-
-test('bulkAction - delete - confirm - failure', async () => {
-  comp.selectedAction = 'delete';
-  comp.showBulkDeleteConfirmDialog = true;
-  comp.selectedCount = 2;
-  comp.selectAllIndeterminate = true;
-  comp.eventData = [{ _isSelected: true, soc_id: "1" }, { _isSelected: false, soc_id: "2" }, { _isSelected: true, soc_id: "3" }];
-  const err = { response: { data: "ERROR_BULK_COMMUNITY" } }
-  const mock = resetPapi().mockPapi('post', null, err);
-
-  await comp.bulkAction(true);
-
-  expect(comp.showBulkDeleteConfirmDialog).toBe(false);
-  expect(comp.selectAllState).toBe(false);
-  expect(comp.selectAllIndeterminate).toBe(true);
-  expect(comp.selectedCount).toBe(2);
-  expect(mock).toHaveBeenCalledTimes(1);
-  expect(mock).toHaveBeenCalledWith('detection/bulk/delete', { ids: ["1", "3"] });
-  expect(comp.$root.error).toBe(true);
-  expect(comp.$root.errorMessage).toBe(comp.i18n.ERROR_BULK_COMMUNITY);
-});
-
-test('reconstructQuery', () => {
-  comp.query = "bar: 'hi' | groupby x"
-  comp.obtainQueryDetails();
-  comp.reconstructQuery();
-  comp.querySearch = "foo: 1"
-
-  // Advanced mode and showFullQuery false so should reconstruct query using new custom filter
-  comp.advanced = true;
-  comp.showFullQuery = false;
-  comp.$router.resolve = jest.fn();
-  comp.$router.resolve.mockReturnValueOnce({ fullPath: 'A' });
-  comp.$router.resolve.mockReturnValueOnce({ fullPath: 'B' });
-  comp.queryModified();
-  expect(comp.query).toBe("foo: 1 | groupby x");
-
-  expect(comp.$router.resolve).toHaveBeenCalledTimes(2);
-});
-
-test('queryModified', () => {
-  comp.$root.loading = true; // prevent any notification logic from running
-  comp.query = "bar: 'hi' | groupby x"
-  comp.obtainQueryDetails();
-  comp.querySearch = "foo: 1"
-
-  // Basic mode, should not reconstruct query
-  comp.queryModified();
-  expect(comp.query).toBe("bar: 'hi' | groupby x");
-
-  // Advanced mode, but showFullQuery is true so should not reconstruct query
-  comp.advanced = true;
-  comp.showFullQuery = true;
-  comp.queryModified();
-  expect(comp.query).toBe("bar: 'hi' | groupby x");
-
-  // Advanced mode and showFullQuery false so should reconstruct query using new custom filter
-  comp.advanced = true;
-  comp.showFullQuery = false;
-  comp.queryModified();
-  expect(comp.query).toBe("foo: 1 | groupby x");
-});
-
-test('getDisplayedQueryVar', () => {
-  expect(comp.getDisplayedQueryVar()).toBe('queryName');
-
-  comp.advanced = true;
-  comp.showFullQuery = true;
-  expect(comp.getDisplayedQueryVar()).toBe('query');
-
-  comp.advanced = true;
-  comp.showFullQuery = false;
-  expect(comp.getDisplayedQueryVar()).toBe('querySearch');
-});
-
-test('bulkUpdateReport - error', () => {
-  let stats = {
-    error: 1,
-  };
-
-  comp.bulkUpdateReport(stats)
-
-  expect(comp.$root.error).toBe(true);
-  expect(comp.$root.errorMessage).toBe('1 error(s) arose during a bulk operation. Please check the SOC logs for more information.');
-});
-
-test('bulkUpdateReport - update success', () => {
-  let stats = {
-    time: 10,
-    filtered: 0,
-    verb: 'update',
-    modified: 2,
-    total: 2,
-  };
-
-  comp.bulkUpdateReport(stats)
-
-  expect(comp.$root.info).toBe(true);
-  expect(comp.$root.infoMessage).toBe('Bulk update successfully updated 2 of 2 events. (10s)');
-});
-
-test('bulkUpdateReport - delete success', () => {
-  let stats = {
-    time: 1000,
-    filtered: 0,
-    verb: 'delete',
-    modified: 200,
-    total: 200,
-  };
-
-  comp.bulkUpdateReport(stats)
-
-  expect(comp.$root.info).toBe(true);
-  expect(comp.$root.infoMessage).toBe('Bulk delete successfully deleted 200 of 200 events. (16m 40s)');
-});
-
-test('bulkUpdateReport - filtered success', () => {
-  let stats = {
-    time: 200,
-    filtered: 1,
-    verb: 'update',
-    modified: 20,
-    total: 20,
-  };
-
-  comp.bulkUpdateReport(stats)
-
-  expect(comp.$root.warning).toBe(true);
-  expect(comp.$root.warningMessage).toBe('Bulk update successfully updated 20 of 20 events. However, the statuses of 1 of the updated detections are controlled by the current regex filter settings and were reverted. <a href="/#/config?s=soc.config.server.modules.suricataengine" data-aid="warning_bulk_update_configure_filters">Click here to configure those filters.</a> (3m 20s)'
-);
-});
-
-test('toggleQuickAction - Tune Detection, Yara => Source Tab, Other Engines => Tuning Tab', async () => {
-  comp.category = 'alerts';
-  comp.escalationMenuVisible = comp.quickActionVisible = false;
-  let event = { "rule.uuid": 'id' }
-  let det = { id: 'onionId', engine: 'elastalert' };
-
-  resetPapi().mockPapi('get', { data: det }, null);
-
-  await comp.toggleQuickAction({}, event, -1, null, null);
-  await new Promise(resolve => setTimeout(resolve, 0)); // Let promises resolve
-  expect(comp.quickActionDetId).toBe('onionId');
-  expect(comp.tuneDetectionTabTarget).toBe('tuning');
-  expect(comp.highlightedDetection)
-
-  resetPapi().mockPapi('get', { data: { id: 'onionId', engine: 'suricata' } }, null);
-  comp.highlightedAlertInfo = comp.highlightedDetection = null;
-
-  await comp.toggleQuickAction({}, event, -1, null, null);
-  await new Promise(resolve => setTimeout(resolve, 0)); // Let promises resolve
-  expect(comp.quickActionDetId).toBe('onionId');
-  expect(comp.tuneDetectionTabTarget).toBe('tuning');
-
-  resetPapi().mockPapi('get', { data: { id: 'onionId', engine: 'strelka' } }, null);
-  comp.highlightedAlertInfo = comp.highlightedDetection = null;
-
-  await comp.toggleQuickAction({}, event, -1, null, null);
-  await new Promise(resolve => setTimeout(resolve, 0)); // Let promises resolve
-  expect(comp.quickActionDetId).toBe('onionId');
-  expect(comp.tuneDetectionTabTarget).toBe('source');
-});
-
-test('buildDetectionEngineHuntQuery', () => {
-  comp.detectionEngineStatusQueries = {
-    elastalert: {
-      default: 'default',
-      IntegrityFailure: 'IntegrityFailure',
-    },
-    suricata: {
-      default: 'default',
-      Healthy: 'Healthy',
-    },
-    strelka: {
-      SyncFailure: 'SyncFailure',
-    }
-  };
-  comp.$root.statusByGridId[''] = {
-    detections: {
-      elastalert: {
-        syncFailure: 1,
-      },
-      suricata: {}, // Healthy
-      strelka: {
-        syncing: 1,
-      },
-    }
-  };
-
-  // miss, fallback to default
-  let query = comp.buildDetectionEngineHuntQuery('elastalert');
-  expect(query).toBe('default');
-
-  // hit
-  query = comp.buildDetectionEngineHuntQuery('suricata');
-  expect(query).toBe('Healthy');
-
-  // miss, no default specified, fallback to simple query
-  query = comp.buildDetectionEngineHuntQuery('strelka');
-  expect(query).toBe(`tags:so-soc AND strelka | groupby log.level | groupby event.action | groupby soc.fields.error`);
-});
-
-test('lookupFieldValue', () => {
-  const record = {
-    'so_foobar.fields.that': 'bar',
-    'this': 10,
-    'that': 'foo',
-    'so_foobar.fields.message': 100,
-  };
-
-  expect(comp.lookupFieldValue(record, 'this')).toBe(10);
-  expect(comp.lookupFieldValue(record, 'that')).toBe('foo');
-  expect(comp.lookupFieldValue(record, 'message')).toBe(100);
-});
-
-test('debounceChartResize', () => {
-  const chart = {
-    options: {
-      responsive: true
-    }
-  };
-  comp.chartResizeTracker = {};
-
-  // Initial call
-  comp.debounceChartResize(chart, { width: 100 });
-  expect(comp.chartResizeTracker[chart].length).toBe(1);
-  expect(comp.chartResizeTracker[chart][0].size.width).toBe(100);
-
-  // Second call with same size
-  comp.debounceChartResize(chart, { width: 100 });
-  expect(comp.chartResizeTracker[chart].length).toBe(2);
-
-  // Simulate flapping
-  for (let i = 0; i < 20; i++) {
-    comp.debounceChartResize(chart, { width: 100 + (i % 2 === 0 ? 1 : -1) });
-  }
-  expect(chart.options.responsive).toBe(false);
-});
-
-test('debounceChartResize - no flapping', () => {
-  const chart = {
-    options: {
-      responsive: true
-    }
-  };
-  comp.chartResizeTracker = {};
-
-  // Initial call
-  comp.debounceChartResize(chart, { width: 100 });
-  expect(comp.chartResizeTracker[chart].length).toBe(1);
-  expect(comp.chartResizeTracker[chart][0].size.width).toBe(100);
-  expect(chart.options.responsive).toBe(true);
-
-  comp.debounceChartResize(chart, { width: 101 });
-  expect(chart.options.responsive).toBe(true);
-
-  comp.debounceChartResize(chart, { width: 101 });
-  expect(chart.options.responsive).toBe(true);
-
-  comp.debounceChartResize(chart, { width: 151 });
-  expect(comp.chartResizeTracker[chart].length).toBe(4);
-  expect(chart.options.responsive).toBe(true);
-});
-
-test('sortBySeverity', () => {
-  const tests = [
-    { A: 'CRITICAL', B: 'CRITICAL', expected: 0 },
-    { A: 'critical', B: 'high', expected: 1 },
-    { A: 'critical', B: 'medium', expected: 2 },
-    { A: 'critical', B: 'low', expected: 3 },
-    { A: 'critical', B: 'informational', expected: 4 },
-    { A: 'critical', B: 'unknown', expected: 5 },
-    { A: 'critical', B: 'foobar', expected: 6 },
-
-    { A: 'high', B: 'high', expected: 0 },
-    { A: 'high', B: 'medium', expected: 1 },
-    { A: 'high', B: 'low', expected: 2 },
-    { A: 'high', B: 'informational', expected: 3 },
-    { A: 'high', B: 'unknown', expected: 4 },
-    { A: 'high', B: 'foobar', expected: 5 },
-
-    { A: 'medium', B: 'medium', expected: 0 },
-    { A: 'medium', B: 'low', expected: 1 },
-    { A: 'medium', B: 'informational', expected: 2 },
-    { A: 'medium', B: 'unknown', expected: 3 },
-    { A: 'medium', B: 'foobar', expected: 4 },
-
-    { A: 'low', B: 'low', expected: 0 },
-    { A: 'low', B: 'informational', expected: 1 },
-    { A: 'low', B: 'unknown', expected: 2},
-    { A: 'low', B: 'foobar', expected: 3 },
-
-    { A: 'unknown', B: 'unknown', expected: 0 },
-    { A: 'unknown', B: 'foobar', expected: 1 },
-
-    { A: 'foobar', B: 'asdf', expected: 0 },
-    { A: 'foobar', B: 'unknown', expected: -1 },
-    { A: 'foobar', B: 'informational', expected: -2 },
-    { A: 'foobar', B: 'low', expected: -3 },
-    { A: 'foobar', B: 'medium', expected: -4 },
-    { A: 'foobar', B: 'high', expected: -5 },
-    { A: 'foobar', B: 'critical', expected: -6 },
-
-    { A: 'critical', B: 'critical', expected: 0 },
-    { A: 'high', B: 'critical', expected: -1 },
-    { A: 'medium', B: 'critical', expected: -2 },
-    { A: 'low', B: 'critical', expected: -3 },
-    { A: 'informational', B: 'critical', expected: -4 },
-    { A: 'unknown', B: 'critical', expected: -5 },
-    { A: 'foobar', B: 'critical', expected: -6 },
-
-    { A: 'high', B: 'high', expected: 0 },
-    { A: 'medium', B: 'high', expected: -1 },
-    { A: 'low', B: 'high', expected: -2 },
-    { A: 'informational', B: 'high', expected: -3 },
-    { A: 'unknown', B: 'high', expected: -4 },
-    { A: 'foobar', B: 'high', expected: -5 },
-
-    { A: 'medium', B: 'medium', expected: 0 },
-    { A: 'low', B: 'medium', expected: -1 },
-    { A: 'informational', B: 'medium', expected: -2 },
-    { A: 'unknown', B: 'medium', expected: -3 },
-    { A: 'foobar', B: 'medium', expected: -4 },
-
-    { A: 'low', B: 'low', expected: 0 },
-    { A: 'informational', B: 'low', expected: -1 },
-    { A: 'unknown', B: 'low', expected: -2},
-    { A: 'foobar', B: 'low', expected: -3 },
-
-    { A: 'unknown', B: 'unknown', expected: 0 },
-    { A: 'foobar', B: 'unknown', expected: -1 },
-
-    { A: 'asdf', B: 'foobar', expected: 0 },
-    { A: 'unknown', B: 'foobar', expected: 1 },
-    { A: 'informational', B: 'foobar', expected: 2 },
-    { A: 'low', B: 'foobar', expected: 3 },
-    { A: 'medium', B: 'foobar', expected: 4 },
-    { A: 'high', B: 'foobar', expected: 5 },
-    { A: 'critical', B: 'foobar', expected: 6 },
-
-    { A: 6, B: 10, expected: 0 },
-  ];
-
-  tests.forEach(({ A, B, expected }) => {
-    const result = comp.sortBySeverity(A, B);
-    expect(result).toBe(expected);
   });
-});
-
-test('countDrilldown', () => {
-  comp.filterRouteDrilldown = null;
-
-  let event = {};
-  event.count = 10;
-
-  comp.countDrilldown(event);
-  expect(comp.$router.length).toBe(0);
-  expect(comp.filterRouteDrilldown).toBe(null);
-
-  event.a = 'a';
-
-  let expected = {
-    path: '',
-    query: {
-      el: 100,
-      filterField: 'a',
-      filterMode: 'DRILLDOWN',
-      filterValue: 'a',
-      gl: 10,
-      q: '',
-      rt: 24,
-      rtu: 'hours',
-      z: ''
-    }
-  };
-
-  comp.countDrilldown(event);
-  expect(comp.$router.length).toBe(1);
-  expect(comp.filterRouteDrilldown).toEqual(expected);
-  expect(comp.$router[0]).toStrictEqual(expected);
-
-  comp.$router = [];
-  comp.filterRouteDrilldown = null;
-
-  event.b = 'b';
-
-  comp.countDrilldown(event);
-  expect(comp.$router.length).toBe(0);
-  expect(comp.filterRouteDrilldown).toBe(null);
-
-  event = {};
-  event.count = 10;
-  event["rule.name"] = 'rule_name';
-  event["event.module"] = 'event_module';
-  event["event.severity_label"] = 'event_severity_label';
-  event["rule.uuid"] = 'rule_uuid';
-
-  expected = {
-    path: "",
-    query: {
-      el: 100,
-      filterField: "rule.name",
-      filterMode: "DRILLDOWN",
-      filterValue: "rule_name",
-      gl: 10,
-      q: "",
-      rt: 24,
-      rtu: "hours",
-      z: ""
-    }
-  };
-
-  comp.countDrilldown(event);
-  expect(comp.$router.length).toBe(1);
-  expect(comp.filterRouteDrilldown).toEqual(expected);
-
-  comp.$router = [];
-  comp.filterRouteDrilldown = null;
-
-  event.a = 10;
-
-  comp.countDrilldown(event);
-  expect(comp.$router.length).toBe(0);
-  expect(comp.filterRouteDrilldown).toBe(null);
-});
-
-test('extractSocValues', () => {
-  let obj = {
-    payload: {
-      a: 1,
-      b: 2,
-      c: 3,
-    },
-    id: 'abc',
-    score: 0.947,
-    type: 'type',
-    timestamp: 'now',
-    source: 'network',
-  };
-
-  const expected = {
-    a: 1,
-    b: 2,
-    c: 3,
-    soc_id: 'abc',
-    soc_score: 0.947,
-    soc_type: 'type',
-    soc_timestamp: 'now',
-    soc_source: 'network',
-  };
-
-  let actual = comp.extractSocValues(obj);
-
-  expect(actual).toStrictEqual(expected);
-});
-
-test('getEventField', () => {
-  let event = {
-    a: 1,
-    timestamp: 'now',
-    'event_data.a': 2,
-    'event_data.b': 3,
-  };
-
-  const a = comp.getEventField(event, 'a');
-  const eventdataA = comp.getEventField(event, 'event_data.a');
-  const b = comp.getEventField(event, 'b');
-  const timestamp = comp.getEventField(event, 'timestamp');
-
-  expect(a).toBe(1);
-  expect(eventdataA).toBe(2);
-  expect(b).toBe('');
-  expect(timestamp).toBe('now');
-});
-
-test('buildQuestionRange', () => {
-  const event = {
-    '@timestamp': '2023-10-01T12:00:00Z',
-  };
-
-  comp.zone = 'America/Denver';
-
-  let range = comp.buildQuestionRange(event, '+/-30d');
-  expect(range).toBe('2023/09/01 06:00:00 AM - 2023/10/31 06:00:00 AM');
-
-  range = comp.buildQuestionRange(event, '-15m');
-  expect(range).toBe('2023/10/01 05:45:00 AM - 2023/10/01 06:00:00 AM');
-
-  range = comp.buildQuestionRange(event, '2h');
-  expect(range).toBe('2023/10/01 06:00:00 AM - 2023/10/01 08:00:00 AM');
-
-  range = comp.buildQuestionRange(event, '+/-60s');
-  expect(range).toBe('2023/10/01 05:59:00 AM - 2023/10/01 06:01:00 AM');
-
-  range = comp.buildQuestionRange(event, '');
-  expect(range).toBe('');
-
-  range = comp.buildQuestionRange(event, 'X');
-  expect(range).toBe('');
-
-  range = comp.buildQuestionRange(event, '10p');
-  expect(range).toBe('');
-});
-
-test('sortAggregateEvents', () => {
-  let tests = [
-    [
-      { expectedPos: 1, value: 2 },
-      { expectedPos: 2, value: 1 },
-      { expectedPos: 0, value: 3 },
-    ],
-    [
-      { expectedPos: 2, value: 40 },
-      { expectedPos: 3, value: 30 },
-      { expectedPos: 0, value: 60 },
-      { value: 10 },
-      { expectedPos: 1, value: 50 },
-      { expectedPos: 4, value: 20 },
-    ],
-    [
-      { expectedPos: 0, value: 0 },
-    ],
-  ];
-
-  for (let events of tests) {
-    const sortedEvents = comp.sortAggregateEvents(events);
-
-    if (events.length >= 5) {
-      expect(sortedEvents.length).toBe(5);
-    } else {
-      expect(sortedEvents.length).toBe(events.length);
-    }
-
-    for (let i = 0; i < sortedEvents.length; i++) {
-      expect(sortedEvents[i].expectedPos).toBe(i);
-    }
-  }
-});
-
-test('askQuestion', async () => {
-  comp.$root.enableReverseLookup = true;
-  comp.zone = "Etc/UTC";
-
-  let question = {};
-  let event = {
-    field: 'present',
-    func: function() {}, // not present
-  };
-
-  await comp.askQuestion(question, event);
-
-  expect(question.answers.length).toBe(1);
-  expect('field' in question.answers[0].payload).toBe(true);
-  expect('func' in question.answers[0].payload).toBe(false);
-
-  const mock1 = mockPapi('get', {
-    data: {
-      metrics: {
-        biggest: [
-          {
-            payload: {
-              name: 'metric event',
-              ip: '1.1.1.1',
-            },
-          },
-        ],
-      },
-    },
-  });
-
-  const mock2 = mockPapi('put', {
-    then: function () { },
-  });
-
-  question = {
-    range: '-30d',
-    filledOQL: 'OQL Query',
-    isAggregate: true,
-  };
-  event = {
-    'soc_timestamp': '2023-10-01T12:00:00Z',
-  };
-
-  await comp.askQuestion(question, event);
-
-  expect(question.answers.length).toBe(1);
-  expect(question.answers[0].payload.name).toBe('metric event');
-  expect('error' in question).toBe(false);
-
-  expect(mock1).toHaveBeenCalledTimes(1);
-  expect(mock1).toHaveBeenCalledWith('events/', {
-    params: {
-      query: 'OQL Query',
-      range: '2023/09/01 12:00:00 PM - 2023/10/01 12:00:00 PM',
-      format: '2006/01/02 3:04:05 PM',
-      zone: 'Etc/UTC',
-      metricLimit: 5,
-      eventLimit: 5,
-    },
-  });
-
-  expect(mock2).toHaveBeenCalledTimes(1);
-  expect(mock2).toHaveBeenCalledWith('util/reverse-lookup', ['1.1.1.1'], {params: {gridId: ''}});
-
-  resetPapi();
-  const mock3 = mockPapi('get', null, new Error('something went wrong'));
-
-  question = {
-    range: '-30d',
-    filledOQL: 'OQL Query',
-    isAggregate: false,
-  };
-  event = {
-    'soc_timestamp': '2023-10-01T12:00:00Z',
-  };
-
-  await comp.askQuestion(question, event);
-
-  expect(question.answers.length).toBe(0);
-  expect('error' in question).toBe(true);
-  expect(question.error).toBe(true);
-
-  expect(mock3).toHaveBeenCalledTimes(1);
-  expect(mock3).toHaveBeenCalledWith('events/', {
-    params: {
-      query: 'OQL Query | sortby @timestamp',
-      range: '2023/09/01 12:00:00 PM - 2023/10/01 12:00:00 PM',
-      format: '2006/01/02 3:04:05 PM',
-      zone: 'Etc/UTC',
-      metricLimit: 5,
-      eventLimit: 5,
-    },
-  });
-
-  comp.$root.enableReverseLookup = false;
-  resetPapi();
-});
-
-test('queryVariableSubstitution - handles simple variable substitution', () => {
-    const event = {
-      'host.name': 'test-host',
-      'source.ip': '192.168.1.1'
-    };
-    const playbooks = [{
-      questions: [{
-        query: 'hostname: {host.name}\nsource_ip: {source.ip}'
-      }]
-    }];
-
-    comp.queryVariableSubstitution(event, playbooks);
-    expect(playbooks[0].questions[0].filledQuery).toBe(
-      'hostname: test-host\nsource_ip: 192.168.1.1'
-    );
-  });
-
-test('queryVariableSubstitution - handles array fields with proper indentation', () => {
-    const event = {
-      'network.private_ip': ['192.168.1.1', '10.0.0.1'],
-      'network.public_ip': ['203.0.113.1']
-    };
-    const playbooks = [{
-      questions: [{
-        query: '    private_ips: {network.private_ip}\n    public_ips: {network.public_ip}'
-      }]
-    }];
-
-    comp.queryVariableSubstitution(event, playbooks);
-    expect(playbooks[0].questions[0].filledQuery).toBe(
-      '    private_ips:\n        - 192.168.1.1\n        - 10.0.0.1\n    public_ips:\n        - 203.0.113.1'
-    );
-  });
-
-test('queryVariableSubstitution - handles missing fields with NODATA', () => {
-    const event = {
-      'host.name': 'test-host'
-    };
-    const playbooks = [{
-      questions: [{
-        query: 'hostname: {host.name}\nip: {missing.field}'
-      }]
-    }];
-
-    comp.queryVariableSubstitution(event, playbooks);
-    expect(playbooks[0].questions[0].filledQuery).toBe(
-      'hostname: test-host\nip: NODATA'
-    );
-  });
-
-test('queryVariableSubstitution - handles array fields with dashes', () => {
-    const event = {
-      'network.private_ip': ['192.168.1.1', '10.0.0.1']
-    };
-    const playbooks = [{
-      questions: [{
-        query: '    - private_ips: {network.private_ip}'
-      }]
-    }];
-
-    comp.queryVariableSubstitution(event, playbooks);
-    expect(playbooks[0].questions[0].filledQuery).toBe(
-      '    - private_ips:\n        - 192.168.1.1\n        - 10.0.0.1'
-    );
-  });
-
-test('queryVariableSubstitution - handles real-world query format', () => {
-  const event = {
-    'network.private_ip': ['192.168.1.1', '10.0.0.1'],
-    'dns.query_name': 'malicious.com',
-    'network.public_ip': ['203.0.113.1'],
-    'related.ip': ['203.0.113.1', '192.168.1.2'],
-  };
-  const playbooks = [{
-    questions: [{
-      query: `aggregation: false
-logsource:
-  category: network
-  service: dns
-detection:
-    selection:
-       - src_ip: '{network.private_ip}'
-       - dns.query.name|contains: '{dns.query_name}'
-       - dns.resolved_ip: '{network.public_ip}'
-    filter:
-      dst_ip: '{related.ip}'
-    condition: selection and filter
-fields:
-    - dns.query.name
-    - dns.query.type_name
-    - dns.resolved_ip
-    - dns.response.code_name`
-    }]
-  }];
-
-  comp.queryVariableSubstitution(event, playbooks);
-  expect(playbooks[0].questions[0].filledQuery).toBe(
-    `aggregation: false
-logsource:
-  category: network
-  service: dns
-detection:
-    selection:
-       - src_ip:
-           - 192.168.1.1
-           - 10.0.0.1
-       - dns.query.name|contains: 'malicious.com'
-       - dns.resolved_ip:
-           - 203.0.113.1
-    filter:
-      dst_ip:
-          - 203.0.113.1
-          - 192.168.1.2
-    condition: selection and filter
-fields:
-    - dns.query.name
-    - dns.query.type_name
-    - dns.resolved_ip
-    - dns.response.code_name`
-  );
-});
-
-test('fetchNewestEvent', async () => {
-  const eventSearch = mockPapi('get', { data: { events: [{ id: '100', payload: { a: 1, b: "2", c: true } }] } });
-  
-  comp.filterToggles = [
-    {
-      name: 'acknowledged',
-      enabled: false,
-    },
-    {
-      name: 'escalated',
-      enabled: false,
-    },
-  ];
-  comp.dateRange = 'x - y';
-  comp.zone = 'zone';
-
-  var item = {
-    a: 1,
-    b: "2",
-    'event_data.c': true,
-    count: 10,
-  };
-
-  await comp.fetchNewestEvent(item);
-
-  expect(item.newest.soc_id).toBe('100');
-  
-  expect(eventSearch).toHaveBeenCalledTimes(1);
-  expect(eventSearch).toHaveBeenCalledWith('events/', {
-    params: {
-      query: 'a:"1" AND b:"2" AND c:"true" AND NOT event.acknowledged:true AND NOT event.escalated:true | sortby @timestamp',
-      range: 'x - y',
-      format: comp.i18n.timePickerSample,
-      zone: 'zone',
-      eventLimit: 1,
-      metricLimit: 0,
-    }
-  });
-
-  comp.filterToggles[0].enabled = true;
-  comp.filterToggles[1].enabled = true;
-  delete item.newest;
-  resetPapi();
-  const eventSearch2 = mockPapi('get', { data: { events: [{ id: '100', payload: { a: 1, b: "2", c: true } }] } });
-
-  await comp.fetchNewestEvent(item);
-
-  expect(item.newest.soc_id).toBe('100');
-  
-  expect(eventSearch2).toHaveBeenCalledTimes(1);
-  expect(eventSearch2).toHaveBeenCalledWith('events/', {
-    params: {
-      query: 'a:"1" AND b:"2" AND c:"true" AND event.acknowledged:true AND event.escalated:true | sortby @timestamp',
-      range: 'x - y',
-      format: comp.i18n.timePickerSample,
-      zone: 'zone',
-      eventLimit: 1,
-      metricLimit: 0,
-    }
-  });
-
-  resetPapi();
-});
-
-test('isComplexQuery', () => {
-  expect(comp.isComplexQuery('')).toBe(false);
-  expect(comp.isComplexQuery('*')).toBe(false);
-  expect(comp.isComplexQuery('foo:bar')).toBe(false);
-  expect(comp.isComplexQuery('foo:"bar"')).toBe(false);
-  expect(comp.isComplexQuery('foo:"bar \"car\""')).toBe(false);
-  expect(comp.isComplexQuery('foo:bar AND cat:dog')).toBe(false);
-  expect(comp.isComplexQuery('foo:bar AND qry:"(cat:dog OR fish:bowl)"')).toBe(false);
-  expect(comp.isComplexQuery('foo:bar AND qry:"this is a \"test\" (cat:dog OR fish:bowl)"')).toBe(false);
-  expect(comp.isComplexQuery('foo:bar AND (cat:dog OR fish:bowl)')).toBe(true);
-  expect(comp.isComplexQuery('foo:bar AND cat:(dog OR fish)')).toBe(true);
-  expect(comp.isComplexQuery('foo:bar AND ((cat:(dog OR fish) OR some:thing))')).toBe(true);
-});
-
-test('toggleAllQuestions', () => {
-  let event = {
-    questions: [{}, {}, {}]
-  };
-  let index = 0;
-  let expand = true;
-  comp.expandedPlaybookQuestions = {};
-
-  // expand
-  comp.toggleAllQuestions(event, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [0, 1, 2] });
-
-  // expand a different index
-  index = 1;
-  comp.toggleAllQuestions(event, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [0, 1, 2], 1: [0, 1, 2] });
-
-  // collapse non-existing index
-  index = 2;
-  expand = false;
-  comp.toggleAllQuestions(event, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [0, 1, 2], 1: [0, 1, 2], 2: [] });
-
-  // collapse existing, expanded index
-  index = 0;
-  comp.toggleAllQuestions(event, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [], 1: [0, 1, 2], 2: [] });
-
-  // expand existing, partially expanded
-  comp.expandedPlaybookQuestions[0] = [1];
-  expand = true;
-  comp.toggleAllQuestions(event, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [1, 0, 2], 1: [0, 1, 2], 2: [] });
-
-  // no changes when expanding an already expanded group, even if out of order
-  comp.toggleAllQuestions(event, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [1, 0, 2], 1: [0, 1, 2], 2: [] });
-
-  // no change when collapsing an already collapsed group
-  index = 2;
-  expand = false;
-  comp.toggleAllQuestions(event, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [1, 0, 2], 1: [0, 1, 2], 2: [] });
-
-  // handle aggregate event
-  let aggEvent = {
-    newest: event,
-  };
-
-  expand = true;
-  comp.toggleAllQuestions(aggEvent, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [1, 0, 2], 1: [0, 1, 2], 2: [0, 1, 2] });
-
-  // handle event that doesn't have playbooks yet
-  const _setTimeout = global.setTimeout;
-  const setTimeoutMock = jest.fn();
-  global.setTimeout = setTimeoutMock;
-
-  let emptyEvent = {};
-  index = 0;
-  expand = true;
-
-  comp.toggleAllQuestions(emptyEvent, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [1, 0, 2], 1: [0, 1, 2], 2: [0, 1, 2] });
-  expect(setTimeoutMock).toHaveBeenCalledTimes(1);
-
-  // handle agg event that doesn't have playbooks yet
-  aggEvent.newest = emptyEvent;
-  setTimeoutMock.mockClear();
-  
-  comp.toggleAllQuestions(aggEvent, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [1, 0, 2], 1: [0, 1, 2], 2: [0, 1, 2] });
-  expect(setTimeoutMock).toHaveBeenCalledTimes(1);
-  
-  // handle event that's loading playbooks
-  emptyEvent.playbookLoading = true;
-  emptyEvent.playbooks = null;
-  setTimeoutMock.mockClear();
-  
-  comp.toggleAllQuestions(emptyEvent, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [1, 0, 2], 1: [0, 1, 2], 2: [0, 1, 2] });
-  expect(setTimeoutMock).toHaveBeenCalledTimes(1);
-
-  // handle agg event that's loading playbooks
-  aggEvent.newest = emptyEvent;
-  setTimeoutMock.mockClear();
-  
-  comp.toggleAllQuestions(aggEvent, index, expand);
-  expect(comp.expandedPlaybookQuestions).toStrictEqual({ 0: [1, 0, 2], 1: [0, 1, 2], 2: [0, 1, 2] });
-  expect(setTimeoutMock).toHaveBeenCalledTimes(1);
-
-  global.setTimeout = _setTimeout;
-});
-
-test('loadPlaybook', async () => {
-  let event = {
-    'rule.uuid': '123',
-  };
-
-  const playbooks = [{ id: '1', questions: [] }, { id: '2', questions: [] }];
-
-  let papiMock = mockPapi('get', { data: playbooks });
-
-  await comp.loadPlaybook(event, 0);
-
-  expect(papiMock).toHaveBeenCalledWith('playbook/detection/123');
-  expect(event.playbooks).toStrictEqual(playbooks);
-  expect(event.playbookLoading).toBe(undefined);
-  expect(event.playbookErr).toBe(false);
-
-  resetPapi();
-  papiMock = mockPapi('get', null, 'error');
-
-  event = {
-    'rule.uuid': '456',
-  };
-
-  comp.loadPlaybook(event, 0);
-
-  expect(papiMock).toHaveBeenCalledWith('playbook/detection/456');
-  expect(papiMock).toHaveBeenCalledTimes(1);
-  expect(event.playbookErr).toBe(true);
-  expect(event.playbooks).toBe(null);
-
-  event = {};
-
-  comp.loadPlaybook(event, 0);
-
-  expect(event.playbookErr).toBe(true);
-
-  papiMock.mockClear();
-
-  event = { playbooks: '' };
-
-  comp.loadPlaybook(event, 0);
-  expect(event.playbookErr).toBe(undefined);
-  expect(event.playbookLoading).toBe(undefined);
-  expect(papiMock).toHaveBeenCalledTimes(0);
-
-  event = { playbookLoading: '' };
-
-  comp.loadPlaybook(event, 0);
-  expect(event.playbooks).toBe(undefined);
-  expect(event.playbookErr).toBe(undefined);
-  expect(papiMock).toHaveBeenCalledTimes(0);
-
-  event = { playbookErr: true };
-
-  comp.loadPlaybook(event, 0);
-  expect(event.playbooks).toBe(undefined);
-  expect(event.playbookLoading).toBe(undefined);
-  expect(papiMock).toHaveBeenCalledTimes(0);
-
-  resetPapi();
-});
-
-test('pickQuestionColor', () => {
-  let question = {
-    error: true,
-  };
-
-  let c = comp.pickQuestionColor(question);
-  expect(c).toBe('has-error');
-
-  question = {
-    error: false,
-  };
-
-  c = comp.pickQuestionColor(question);
-  expect(c).toBe('no-data');
-
-  question = {}
-
-  c = comp.pickQuestionColor(question);
-  expect(c).toBe('no-data');
-
-  question = {
-    answers: [],
-  };
-
-  c = comp.pickQuestionColor(question);
-  expect(c).toBe('no-data');
-
-  question = {
-    answers: [{}],
-  };
-
-  c = comp.pickQuestionColor(question);
-  expect(c).toBe('has-answers');
-
-  question = {
-    answers: [{}],
-    error: false,
-  };
-
-  c = comp.pickQuestionColor(question);
-  expect(c).toBe('has-answers');
-
-  question = {
-    answers: [{}],
-    error: true,
-  };
-
-  c = comp.pickQuestionColor(question);
-  expect(c).toBe('has-error');
 });
