+++ html/js/routes/hunt/data.js
@@ -1,0 +1,154 @@
+export default {
+  i18n: {}, // Placeholder for this.$root.i18n
+  params: null,
+  category: '',
+  advanced: false,
+  queryAltered: false,
+  query: '',
+  querySearch: '',
+  queryRemainder: '',
+  queries: [],
+  queryBaseFilter: "",
+  queryName: '',
+  queryFilters: [],
+  queryGroupBys: [],
+  queryGroupByOptions: [],
+  querySortBys: [],
+  queryTableFields: [],
+  queryTableOptions: [],
+  eventFields: {},
+  dateRange: '',
+  relativeTimeEnabled: true,
+  relativeTimeValue: 24,
+  relativeTimeUnit: 30, // RELATIVE_TIME_HOURS
+  relativeTimeUnits: [],
+  autoRefreshInterval: 0,
+  autoRefreshIntervals: [],
+  autoRefreshTimer: null,
+  loaded: false,
+  chartHeight: 200,
+  zone: '',
+  huntPending: false,
+  ackEnabled: false,
+  escalateEnabled: false,
+  viewEnabled: false,
+  createLink: '',
+  collapsedSections: [],
+  filterToggles: [],
+  timelineChartOptions: {},
+  timelineChartData: {},
+  metricsEnabled: false,
+  eventsEnabled: true,
+  topChartOptions: {},
+  topChartData: {},
+  bottomChartOptions: {},
+  bottomChartData: {},
+  groupBys: [],
+  groupByLimitOptions: [10,25,50,100,200,500],
+  groupByLimit: 10,
+  groupByFilter: '',
+  groupByItemsPerPage: 10,
+  groupByFooters: [10,25,50,100,200,500],
+  groupByPage: 1,
+  groupBySortBy: 'count',
+  groupBySortDesc: true,
+  chartLabelMaxLength: 30,
+  chartLabelOtherLimit: 10,
+  chartLabelFieldSeparator: ', ',
+  eventLimitOptions: [10,25,50,100,200,500,1000,2000,5000],
+  eventLimit: 100,
+  eventData: [],
+  eventFilter: '',
+  eventHeaders: [],
+  eventPage: 1,
+  sortBy: [{ key: 'soc_timestamp', order: 'desc' }],
+  sortDesc: true,
+  itemsPerPage: 10,
+  itemsPerPageOptions: [10,25,50,100,200,500,1000],
+  expandedHeaders: [
+    { title: "key", value: "key" },
+    { title: "value", value: "value" }
+  ],
+  totalEvents: 0,
+  fetchTimeSecs: 0,
+  roundTripTimeSecs: 0,
+  mruQueries: [],
+  mruCases: [],
+  autohunt: true,
+  showFullQuery: true,
+  filterRouteInclude: "",
+  filterRouteExclude: "",
+  filterRouteExact: "",
+  filterRouteDrilldown: "",
+  filterRouteLessThan: "",
+  filterRouteLessThanEqual: "",
+  filterRouteGreaterThan: "",
+  filterRouteGreaterThanEqual: "",
+  groupByRoute: "",
+  groupByNewRoute: "",
+  quickActionVisible: false,
+  quickActionTarget: null,
+  quickActionEvent: null,
+  quickActionField: "",
+  quickActionValue: "",
+  quickActionIsNumeric: false,
+  escalationMenuVisible: false,
+  escalationMenuTarget: null,
+  escalationItem: null,
+  escalationGroupIdx: -1,
+  escalateRelatedEventsEnabled: false,
+  aggregationActionsEnabled: false,
+  actions: [],
+  safeStringMaxLength: Number.MAX_SAFE_INTEGER,
+  betweenDialogVisible: false,
+  betweenStart: '',
+  betweenStartEquals: false,
+  betweenEnd: '',
+  betweenEndEquals: false,
+  betweenError: '',
+  equalityOperators: [
+    { title: '<', value: false },
+    { title: '≤', value: true }
+  ],
+  addToCaseDialogVisible: false,
+  mruCases: [],
+  selectedMruCase: null,
+  disableRouteLoad: false,
+  selectAllState: false,
+  selectAllIndeterminate: false,
+  selectedCount: 0,
+  selectedAction: 'enable',
+  bulkActions: [
+    { title: 'Enable', value: 'enable' },
+    { title: 'Disable', value: 'disable' },
+    { title: 'Delete', value: 'delete' },
+  ],
+  quickActionDetId: null,
+  presets: {},
+  manualSyncTargetEngine: null,
+  showBulkDeleteConfirmDialog: false,
+  tuneDetectionTabTarget: null,
+  eventCurrentItems: [],
+  detectionEngineStatusQueries: {},
+  highlightedDetection: null,
+  highlightedAlertInfo: null,
+  showDetailsPanel: true,
+  openPanel: [0],
+  quickActionsOpen: [],
+  escalatingDetectionEvents: false,
+  showAckManyDialog: false,
+  ackManyArgs: [],
+  ackManyVerb: '',
+  menuScrollPos: 0,
+  maxEscalate: 100,
+  chartResizeTracker: {},
+  gridId: null,
+  activeTabs: {},
+  expandedEvents: [],
+  eventColumnWidth: 0,
+  expandedPlaybookQuestions: {},
+
+  // AI Investigation tracking
+  aiInvestigations: {}, // Maps alert ID to investigation results and chat session IDs
+  aiInvestigatedFilter: false, // Client-side filter for AI investigated alerts
+}
