+++ html/pages/alerts.html
@@ -126,6 +126,7 @@
             @acknowledge="acknowledgeAlert"
             @escalate="escalateAlert"
             @dismiss="dismissAlert"
+            @click="showAlertDetails"
           ></alert-card>
         </v-col>
       </v-row>
@@ -184,6 +185,145 @@
           :loading="actionLoading"
         >
           Confirm
+        </v-btn>
+      </v-card-actions>
+    </v-card>
+  </v-dialog>
+
+  <!-- Alert Details Dialog -->
+  <v-dialog v-model="detailsDialog" max-width="900" scrollable>
+    <v-card>
+      <v-card-title class="d-flex align-center">
+        <span>Alert Details</span>
+        <v-spacer></v-spacer>
+        <v-btn
+          icon
+          variant="text"
+          @click="detailsDialog = false"
+        >
+          <v-icon>fa-xmark</v-icon>
+        </v-btn>
+      </v-card-title>
+      <v-divider></v-divider>
+      <v-card-text v-if="selectedAlertDetails">
+        <!-- Alert Header Info -->
+        <v-row class="mb-4">
+          <v-col cols="12">
+            <div class="d-flex align-center mb-2">
+              <v-chip
+                :color="getSeverityColor(selectedAlertDetails.severityLabel)"
+                size="small"
+                class="mr-2"
+              >
+                {{ selectedAlertDetails.severityLabel }}
+              </v-chip>
+              <v-chip
+                v-if="selectedAlertDetails.acknowledged"
+                color="info"
+                variant="outlined"
+                size="small"
+                class="mr-2"
+              >
+                <v-icon start size="x-small">fa-check</v-icon>
+                Acknowledged
+              </v-chip>
+              <v-chip
+                v-if="selectedAlertDetails.escalated"
+                color="warning"
+                variant="outlined"
+                size="small"
+                class="mr-2"
+              >
+                <v-icon start size="x-small">fa-briefcase</v-icon>
+                Escalated
+              </v-chip>
+            </div>
+            <h3 class="text-h6">{{ selectedAlertDetails.ruleName }}</h3>
+            <div class="text-caption text--secondary">
+              <v-icon size="x-small">fa-clock</v-icon>
+              {{ new Date(selectedAlertDetails.timestamp).toLocaleString() }}
+            </div>
+          </v-col>
+        </v-row>
+
+        <!-- Network Information -->
+        <v-row class="mb-4">
+          <v-col cols="12" md="6">
+            <v-card variant="outlined">
+              <v-card-title class="text-subtitle-2">Source</v-card-title>
+              <v-card-text>
+                <div><strong>IP:</strong> {{ selectedAlertDetails.sourceIp || 'N/A' }}</div>
+                <div><strong>Port:</strong> {{ selectedAlertDetails.sourcePort || 'N/A' }}</div>
+              </v-card-text>
+            </v-card>
+          </v-col>
+          <v-col cols="12" md="6">
+            <v-card variant="outlined">
+              <v-card-title class="text-subtitle-2">Destination</v-card-title>
+              <v-card-text>
+                <div><strong>IP:</strong> {{ selectedAlertDetails.destIp || 'N/A' }}</div>
+                <div><strong>Port:</strong> {{ selectedAlertDetails.destPort || 'N/A' }}</div>
+              </v-card-text>
+            </v-card>
+          </v-col>
+        </v-row>
+
+        <!-- Alert Metadata -->
+        <v-row class="mb-4">
+          <v-col cols="12">
+            <v-card variant="outlined">
+              <v-card-title class="text-subtitle-2">Alert Information</v-card-title>
+              <v-card-text>
+                <v-row>
+                  <v-col cols="12" md="6">
+                    <div><strong>Alert ID:</strong> {{ selectedAlertDetails.id }}</div>
+                    <div><strong>Module:</strong> {{ selectedAlertDetails.module || 'N/A' }}</div>
+                    <div><strong>Category:</strong> {{ selectedAlertDetails.category || 'N/A' }}</div>
+                  </v-col>
+                  <v-col cols="12" md="6">
+                    <div><strong>Rule UUID:</strong> {{ selectedAlertDetails.ruleUuid || 'N/A' }}</div>
+                    <div><strong>Severity:</strong> {{ selectedAlertDetails.severity }} ({{ selectedAlertDetails.severityLabel }})</div>
+                  </v-col>
+                </v-row>
+              </v-card-text>
+            </v-card>
+          </v-col>
+        </v-row>
+
+        <!-- All Fields (Raw Data) -->
+        <v-row>
+          <v-col cols="12">
+            <v-expansion-panels>
+              <v-expansion-panel>
+                <v-expansion-panel-title>
+                  <v-icon class="mr-2">fa-code</v-icon>
+                  All Event Fields
+                </v-expansion-panel-title>
+                <v-expansion-panel-text>
+                  <pre style="overflow-x: auto; font-size: 12px;">{{ JSON.stringify(selectedAlertDetails.rawData, null, 2) }}</pre>
+                </v-expansion-panel-text>
+              </v-expansion-panel>
+            </v-expansion-panels>
+          </v-col>
+        </v-row>
+      </v-card-text>
+      <v-divider></v-divider>
+      <v-card-actions>
+        <v-spacer></v-spacer>
+        <v-btn
+          color="primary"
+          variant="text"
+          :href="getHuntUrl(selectedAlertDetails)"
+          target="_blank"
+        >
+          <v-icon start>fa-magnifying-glass</v-icon>
+          View in Hunt
+        </v-btn>
+        <v-btn
+          variant="text"
+          @click="detailsDialog = false"
+        >
+          Close
         </v-btn>
       </v-card-actions>
     </v-card>
